refact: tear up large file integration.rs into module
