--- filter-repo-rs/tests/integration.rs
@@ -1,5592 +1,0 @@
-use std::env;
-use std::fs::{self, File};
-use std::io::{Read, Write};
-use std::path::{Path, PathBuf};
-use std::process::Command;
-
-use filter_repo_rs as fr;
-use regex::bytes::Regex;
-
-fn mktemp(prefix: &str) -> PathBuf {
-    // Place temp repos under target/ to avoid Windows safe.directory issues
-    let mut p = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    p.push("target");
-    p.push("it");
-    static COUNT: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
-    let pid = std::process::id();
-    let t = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_nanos();
-    let c = COUNT.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    p.push(format!("{}_{}_{}_{}", prefix, pid, t, c));
-    p
-}
-
-fn run_git(dir: &Path, args: &[&str]) -> (i32, String, String) {
-    let out = Command::new("git")
-        .current_dir(dir)
-        .args(args)
-        .output()
-        .expect("run git");
-    let code = out.status.code().unwrap_or(-1);
-    let stdout = String::from_utf8_lossy(&out.stdout).to_string();
-    let stderr = String::from_utf8_lossy(&out.stderr).to_string();
-    (code, stdout, stderr)
-}
-
-fn write_file(dir: &Path, rel: &str, contents: &str) {
-    let path = dir.join(rel);
-    if let Some(p) = path.parent() {
-        fs::create_dir_all(p).unwrap();
-    }
-    let mut f = File::create(&path).unwrap();
-    f.write_all(contents.as_bytes()).unwrap();
-}
-
-fn init_repo() -> PathBuf {
-    let repo = mktemp("fr_rs_it");
-    fs::create_dir_all(&repo).unwrap();
-    // plain init (avoid optional flags for maximum compatibility)
-    let (c, _o, e) = run_git(&repo, &["init"]);
-    assert_eq!(c, 0, "git init failed: {}", e);
-    run_git(&repo, &["config", "user.name", "A U Thor"]).0;
-    run_git(&repo, &["config", "user.email", "a.u.thor@example.com"]).0;
-    write_file(&repo, "README.md", "hello");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "init commit"]).0, 0);
-    repo
-}
-
-fn run_tool(dir: &Path, configure: impl FnOnce(&mut fr::Options)) -> (i32, String, String) {
-    let mut opts = fr::Options::default();
-    opts.source = dir.to_path_buf();
-    opts.target = dir.to_path_buf();
-    configure(&mut opts);
-    let res = fr::run(&opts);
-    let code = if res.is_ok() { 0 } else { 1 };
-    (code, String::new(), String::new())
-}
-
-fn current_branch(repo: &Path) -> String {
-    let (_, branch, _) = run_git(repo, &["symbolic-ref", "--short", "HEAD"]);
-    let mut branch = branch.trim().to_string();
-    if branch.is_empty() || branch == "HEAD" {
-        let (_, alt, _) = run_git(repo, &["rev-parse", "--abbrev-ref", "HEAD"]);
-        branch = alt.trim().to_string();
-    }
-    branch
-}
-
-#[test]
-fn analyze_mode_produces_human_report() {
-    let repo = init_repo();
-    let mut opts = fr::Options::default();
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.mode = fr::Mode::Analyze;
-    let report = fr::analysis::generate_report(&opts).expect("generate analysis report");
-    assert!(
-        report.metrics.refs_total >= 1,
-        "expected refs to be counted"
-    );
-    assert!(
-        !report.warnings.is_empty(),
-        "expected at least one informational warning"
-    );
-    fr::analysis::run(&opts).expect("analyze mode should render without error");
-}
-
-#[test]
-fn analyze_mode_emits_json() {
-    let repo = init_repo();
-    let mut opts = fr::Options::default();
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.mode = fr::Mode::Analyze;
-    let report = fr::analysis::generate_report(&opts).expect("generate analysis report");
-    let json = serde_json::to_string(&report).expect("serialize report");
-    let v: serde_json::Value = serde_json::from_str(&json).expect("valid json");
-    assert!(
-        v.get("metrics").is_some(),
-        "metrics missing in json: {}",
-        json
-    );
-    assert!(
-        v.get("warnings").is_some(),
-        "warnings missing in json: {}",
-        json
-    );
-    opts.analyze.json = true;
-    fr::analysis::run(&opts).expect("json analyze run should succeed");
-}
-
-#[test]
-fn analyze_mode_limits_top_entries_and_populates_paths() {
-    let repo = init_repo();
-    // create blobs of various sizes so the top list can be truncated
-    for i in 0..5 {
-        let size = (i + 1) * 1024;
-        let contents = "x".repeat(size);
-        write_file(&repo, &format!("data/blob{}.bin", i), &contents);
-    }
-    // create multiple duplicate blobs with distinct contents to ensure truncation
-    for (idx, paths) in [
-        ("A", vec!["dups/a1.txt", "dups/a2.txt", "dups/a3.txt"]),
-        ("B", vec!["dups/b1.txt", "dups/b2.txt"]),
-        ("C", vec!["dups/c1.txt", "dups/c2.txt"]),
-    ] {
-        let payload = format!("duplicate payload {}", idx);
-        for path in paths {
-            write_file(&repo, path, &payload);
-        }
-    }
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "populate blobs"]).0, 0);
-
-    let mut opts = fr::Options::default();
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.mode = fr::Mode::Analyze;
-    opts.analyze.top = 2;
-    opts.analyze.thresholds.warn_blob_bytes = 1500;
-    let report = fr::analysis::generate_report(&opts).expect("generate analysis report");
-
-    assert!(
-        report.metrics.largest_blobs.len() <= opts.analyze.top,
-        "largest blobs exceeded top limit"
-    );
-    assert!(
-        report.metrics.blobs_over_threshold.len() <= opts.analyze.top,
-        "threshold hits exceeded top limit"
-    );
-    assert!(
-        report.metrics.duplicate_blobs.len() <= opts.analyze.top,
-        "duplicate blob list exceeded top limit"
-    );
-    assert!(
-        report
-            .metrics
-            .largest_blobs
-            .iter()
-            .all(|b| b.path.is_some()),
-        "expected sample paths for top blobs"
-    );
-    assert!(
-        report
-            .metrics
-            .blobs_over_threshold
-            .iter()
-            .all(|b| b.path.is_some()),
-        "expected sample paths for threshold hits"
-    );
-    assert!(
-        report
-            .metrics
-            .duplicate_blobs
-            .iter()
-            .all(|d| d.example_path.is_some()),
-        "expected example paths for duplicates"
-    );
-}
-
-#[test]
-fn analyze_mode_warns_on_commit_thresholds() {
-    let repo = init_repo();
-    // oversized commit message that should exceed the configured threshold
-    write_file(&repo, "logs.txt", &"L".repeat(64));
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", &"M".repeat(64)]).0, 0);
-    let (_, long_oid, _) = run_git(&repo, &["rev-parse", "HEAD"]);
-    let long_oid = long_oid.trim().to_string();
-    // determine the name of the default branch (e.g. master or main)
-    // prefer symbolic-ref, but fall back to rev-parse if needed
-    let (_, base_branch, _) = run_git(&repo, &["symbolic-ref", "--short", "HEAD"]);
-    let mut base_branch = base_branch.trim().to_string();
-    if base_branch.is_empty() || base_branch == "HEAD" {
-        let (_, alt, _) = run_git(&repo, &["rev-parse", "--abbrev-ref", "HEAD"]);
-        base_branch = alt.trim().to_string();
-    }
-
-    // create a feature branch and diverging history to produce a merge commit
-    assert_eq!(run_git(&repo, &["checkout", "-b", "feature"]).0, 0);
-    write_file(&repo, "feature.txt", "feature work");
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "feature commit"]).0, 0);
-
-    // return to the original default branch, regardless of its name
-    assert_eq!(run_git(&repo, &["checkout", &base_branch]).0, 0);
-    write_file(&repo, "master.txt", "master work");
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "master commit"]).0, 0);
-
-    let merge_msg = "Merge branch 'feature' with an explanation that exceeds the warn threshold";
-    assert_eq!(run_git(&repo, &["merge", "feature", "-m", merge_msg]).0, 0);
-
-    let mut opts = fr::Options::default();
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.mode = fr::Mode::Analyze;
-    opts.analyze.thresholds.warn_commit_msg_bytes = 32;
-    opts.analyze.thresholds.warn_max_parents = 1;
-
-    let report = fr::analysis::generate_report(&opts).expect("generate analysis report");
-
-    assert!(
-        report.metrics.max_commit_parents > 1,
-        "expected merge commit to exceed parent threshold"
-    );
-    assert!(
-        report
-            .metrics
-            .oversized_commit_messages
-            .iter()
-            .any(|m| m.oid.trim() == long_oid),
-        "expected long commit message to be recorded"
-    );
-    assert!(
-        report
-            .warnings
-            .iter()
-            .any(|w| w.message.contains(&long_oid)),
-        "expected warning mentioning oversized commit message"
-    );
-    assert!(
-        report
-            .warnings
-            .iter()
-            .any(|w| w.message.contains("parents")),
-        "expected warning about excessive commit parents"
-    );
-}
-
-#[test]
-fn merge_parents_dedup_when_side_branch_pruned() {
-    let repo = init_repo();
-    let base_branch = current_branch(&repo);
-
-    write_file(&repo, "keep.txt", "base");
-    assert_eq!(run_git(&repo, &["add", "keep.txt"]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "add keep path"]).0, 0);
-
-    assert_eq!(run_git(&repo, &["checkout", "-b", "feature-branch"]).0, 0);
-    write_file(&repo, "drop.txt", "side change");
-    assert_eq!(run_git(&repo, &["add", "drop.txt"]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "side branch change"]).0, 0);
-
-    assert_eq!(run_git(&repo, &["checkout", &base_branch]).0, 0);
-    assert_eq!(
-        run_git(&repo, &["merge", "--no-ff", "--no-commit", "feature-branch"]).0,
-        0,
-        "expected merge to succeed",
-    );
-    write_file(&repo, "keep.txt", "merge resolution");
-    assert_eq!(run_git(&repo, &["add", "keep.txt"]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "merge feature branch"]).0, 0);
-
-    let (code, _, _) = run_tool(&repo, |opts| {
-        opts.paths.push(b"keep.txt".to_vec());
-    });
-    assert_eq!(code, 0, "filter run should succeed");
-
-    let filtered_path = repo
-        .join(".git")
-        .join("filter-repo")
-        .join("fast-export.filtered");
-    let data = fs::read(&filtered_path).expect("read filtered fast-export");
-    let marker = b"merge feature branch";
-    let pos = data
-        .windows(marker.len())
-        .position(|w| w == marker)
-        .expect("merge commit message present");
-    let commit_tag = b"\ncommit ";
-    let commit_start = data[..pos]
-        .windows(commit_tag.len())
-        .rposition(|w| w == commit_tag)
-        .map(|idx| idx + 1)
-        .unwrap_or(0);
-
-    let mut idx = commit_start;
-    let mut parents: Vec<Vec<u8>> = Vec::new();
-    while idx < data.len() {
-        let line_end = match data[idx..].iter().position(|&b| b == b'\n') {
-            Some(end) => idx + end,
-            None => break,
-        };
-        let line = &data[idx..line_end];
-        idx = line_end + 1;
-        if line.is_empty() {
-            break;
-        }
-        if line.starts_with(b"data ") {
-            let len = std::str::from_utf8(&line[b"data ".len()..])
-                .expect("utf8 data header")
-                .trim()
-                .parse::<usize>()
-                .expect("numeric data length");
-            idx = idx.saturating_add(len);
-            if idx < data.len() && data[idx] == b'\n' {
-                idx += 1;
-            }
-            continue;
-        }
-        if line.starts_with(b"from ") || line.starts_with(b"merge ") {
-            parents.push(line.to_vec());
-        }
-    }
-
-    assert_eq!(parents.len(), 1, "expected single parent line: {:?}", parents);
-    assert!(
-        parents[0].starts_with(b"from :"),
-        "parent line should be from-mark: {:?}",
-        parents
-    );
-}
-
-#[test]
-fn tag_rename_lightweight_creates_new_and_deletes_old() {
-    let repo = init_repo();
-    // create lightweight tag
-    assert_eq!(run_git(&repo, &["tag", "v1.0"]).0, 0);
-    // run rename
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-        o.tag_rename = Some((b"v".to_vec(), b"release-".to_vec()));
-    });
-    // verify new exists
-    let (_c2, out, _e2) = run_git(&repo, &["show-ref", "--tags"]);
-    assert!(
-        out.contains("refs/tags/release-1.0"),
-        "expected release-1.0 in tags, got: {}",
-        out
-    );
-    assert!(
-        !out.contains("refs/tags/v1.0"),
-        "old tag v1.0 should be deleted, got: {}",
-        out
-    );
-}
-
-#[test]
-fn tag_rename_annotated_produces_tag_object() {
-    let repo = init_repo();
-    // annotated tag
-    assert_eq!(
-        run_git(&repo, &["tag", "-a", "-m", "hello tag", "v1.0"]).0,
-        0
-    );
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-        o.tag_rename = Some((b"v".to_vec(), b"release-".to_vec()));
-    });
-    // resolve new tag object and check type
-    let (_c1, oid, _e1) = run_git(&repo, &["rev-parse", "refs/tags/release-1.0"]);
-    let oid = oid.trim();
-    let (_c2, typ, _e2) = run_git(&repo, &["cat-file", "-t", oid]);
-    assert_eq!(
-        typ.trim(),
-        "tag",
-        "expected annotated tag object, got type {} for {}",
-        typ,
-        oid
-    );
-}
-
-#[test]
-fn replace_message_edits_commit_and_tag_messages() {
-    let repo = init_repo();
-    // second commit with token 'FOO'
-    write_file(&repo, "src/a.txt", "x");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "commit with FOO token"]).0,
-        0
-    );
-    // annotated tag with token 'FOO'
-    assert_eq!(
-        run_git(&repo, &["tag", "-a", "-m", "tag msg FOO", "v2.0"]).0,
-        0
-    );
-    // replacement file
-    let repl = repo.join("repl.txt");
-    fs::write(&repl, "FOO==>BAR\n").unwrap();
-    // run tool
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.replace_message_file = Some(repl.clone());
-        o.no_data = true;
-    });
-    // check HEAD message
-    let (_c1, msg, _e1) = run_git(&repo, &["log", "-1", "--format=%B"]);
-    assert!(
-        msg.contains("BAR"),
-        "expected commit message to contain BAR, got: {}",
-        msg
-    );
-    assert!(
-        !msg.contains("FOO"),
-        "commit message should be rewritten, got: {}",
-        msg
-    );
-    // check tag message
-    let (_c2, tag_oid, _e2) = run_git(&repo, &["rev-parse", "refs/tags/v2.0"]);
-    let tag_oid = tag_oid.trim();
-    let (_c3, tag_obj, _e3) = run_git(&repo, &["cat-file", "-p", tag_oid]);
-    assert!(
-        tag_obj.contains("BAR"),
-        "expected tag message to contain BAR, got: {}",
-        tag_obj
-    );
-}
-
-#[test]
-fn second_run_rewrites_short_hashes_in_messages() {
-    let repo = init_repo();
-    // First commit touching keep/ and drop/ so path filtering rewrites history
-    write_file(&repo, "keep/data.txt", "first keep\n");
-    write_file(&repo, "drop/data.txt", "first drop\n");
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "seed data"]).0, 0);
-
-    let (_c_old, commit1_full, _e_old) = run_git(&repo, &["rev-parse", "HEAD"]);
-    let commit1_full = commit1_full.trim().to_string();
-    assert_eq!(commit1_full.len(), 40, "expected full commit id");
-    let old_short = commit1_full[..7].to_string();
-
-    // Second commit referencing the short id of the first commit
-    write_file(
-        &repo,
-        "keep/data.txt",
-        &format!("updated keep referencing {}\n", old_short),
-    );
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(
-        run_git(
-            &repo,
-            &["commit", "-m", &format!("mention old short {}", old_short)]
-        )
-        .0,
-        0
-    );
-
-    assert_eq!(
-        run_git(
-            &repo,
-            &["tag", "-a", "-m", &format!("tag cites {}", old_short), "v-short"]
-        )
-        .0,
-        0
-    );
-
-    // First run: filter to keep/ (drop/ removed) to rewrite commits and build commit-map
-    let (code1, _, _) = run_tool(&repo, |o| {
-        o.paths.push(b"keep".to_vec());
-    });
-    assert_eq!(code1, 0, "initial filtering run should succeed");
-
-    let commit_map_path = repo.join(".git").join("filter-repo").join("commit-map");
-    let map_contents = fs::read_to_string(&commit_map_path).expect("read commit-map after first run");
-    let mut new_full: Option<String> = None;
-    for line in map_contents.lines() {
-        let mut parts = line.split_whitespace();
-        if let (Some(old), Some(new)) = (parts.next(), parts.next()) {
-            if old == commit1_full {
-                new_full = Some(new.to_string());
-                break;
-            }
-        }
-    }
-    let new_full = new_full.expect("commit-map should contain first commit mapping");
-    assert_eq!(new_full.len(), 40, "expected mapped commit id length");
-    let new_short = new_full[..7].to_string();
-    assert_ne!(
-        new_short, old_short,
-        "filtering should produce a different commit id to exercise rewriting"
-    );
-
-    // Second run should reuse commit-map to rewrite short hashes inside messages
-    let (code2, _, _) = run_tool(&repo, |o| {
-        o.paths.push(b"keep".to_vec());
-    });
-    assert_eq!(code2, 0, "second filtering run should succeed");
-
-    let (_c_msg, msg, _e_msg) = run_git(&repo, &["log", "-1", "--format=%B"]);
-    assert!(
-        msg.contains(&new_short),
-        "expected rewritten commit message to contain new short {}, got: {}",
-        new_short,
-        msg
-    );
-    assert!(
-        !msg.contains(&old_short),
-        "commit message should no longer contain old short {}, got: {}",
-        old_short,
-        msg
-    );
-
-    let (_c_tag, tag_obj, _e_tag) = run_git(&repo, &["cat-file", "-p", "refs/tags/v-short"]);
-    assert!(
-        tag_obj.contains(&new_short),
-        "expected tag message to contain new short {}, got: {}",
-        new_short,
-        tag_obj
-    );
-    assert!(
-        !tag_obj.contains(&old_short),
-        "tag message should not contain old short {}, got: {}",
-        old_short,
-        tag_obj
-    );
-}
-
-#[test]
-fn writes_commit_map_and_ref_map() {
-    let repo = init_repo();
-    // annotated tag for ref-map
-    run_git(&repo, &["tag", "-a", "-m", "msg", "v3.0"]);
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-        o.tag_rename = Some((b"v".to_vec(), b"release-".to_vec()));
-    });
-    let debug = repo.join(".git").join("filter-repo");
-    let commit_map = debug.join("commit-map");
-    let ref_map = debug.join("ref-map");
-    assert!(
-        commit_map.exists(),
-        "commit-map should exist at {:?}",
-        commit_map
-    );
-    // commit-map should be non-empty
-    let mut s = String::new();
-    File::open(&commit_map)
-        .unwrap()
-        .read_to_string(&mut s)
-        .unwrap();
-    assert!(!s.trim().is_empty(), "commit-map should have content");
-    // ref-map should contain tag rename
-    let mut r = String::new();
-    File::open(&ref_map)
-        .unwrap()
-        .read_to_string(&mut r)
-        .unwrap();
-    assert!(
-        r.contains("refs/tags/v3.0 refs/tags/release-3.0"),
-        "ref-map expected v3.0->release-3.0, got: {}",
-        r
-    );
-}
-
-#[test]
-fn commit_map_records_pruned_commit_as_null() {
-    let repo = init_repo();
-    // commit touching a kept path
-    write_file(&repo, "keep/keep.txt", "keep one");
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "add keep file"]).0, 0);
-
-    // commit touching only a path that will be filtered out
-    write_file(&repo, "drop/drop.txt", "drop me");
-    assert_eq!(run_git(&repo, &["add", "."]).0, 0);
-    assert_eq!(run_git(&repo, &["commit", "-m", "add drop file"]).0, 0);
-
-    let (_code, drop_oid, _e) = run_git(&repo, &["rev-parse", "HEAD"]);
-    let drop_oid = drop_oid.trim().to_string();
-
-    let (code, _o, _e) = run_tool(&repo, |o| {
-        o.paths.push(b"keep".to_vec());
-    });
-    assert_eq!(code, 0, "filter-repo-rs run should succeed");
-
-    let debug_dir = repo.join(".git").join("filter-repo");
-    let commit_map = debug_dir.join("commit-map");
-    assert!(commit_map.exists(), "commit-map should exist after filtering");
-
-    let mut contents = String::new();
-    File::open(&commit_map)
-        .unwrap()
-        .read_to_string(&mut contents)
-        .unwrap();
-    let null_oid = "0000000000000000000000000000000000000000";
-    assert!(
-        contents.contains(&format!("{} {}", drop_oid, null_oid)),
-        "expected commit-map to record pruned commit {}, contents: {}",
-        drop_oid,
-        contents
-    );
-}
-
-fn find_bundles_in(dir: &Path) -> Vec<PathBuf> {
-    fs::read_dir(dir)
-        .expect("failed to read backup directory")
-        .filter_map(|entry| {
-            let entry = entry.ok()?;
-            let path = entry.path();
-            if path.extension().and_then(|ext| ext.to_str()) == Some("bundle") {
-                Some(path)
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
-#[test]
-fn backup_creates_bundle_in_filter_repo_directory() {
-    let repo = init_repo();
-    let (code, _o, _e) = run_tool(&repo, |o| {
-        o.backup = true;
-        o.no_data = true;
-    });
-    assert_eq!(code, 0, "filter-repo-rs run should succeed");
-
-    let backup_dir = repo.join(".git").join("filter-repo");
-    assert!(
-        backup_dir.exists(),
-        "backup directory should exist at {:?}",
-        backup_dir
-    );
-    let bundles = find_bundles_in(&backup_dir);
-    assert!(
-        !bundles.is_empty(),
-        "expected at least one bundle in {:?}, entries: {:?}",
-        backup_dir,
-        bundles
-    );
-}
-
-#[test]
-fn backup_respects_directory_override() {
-    let repo = init_repo();
-    let custom_dir = PathBuf::from("custom-backups");
-    let (code, _o, _e) = run_tool(&repo, |o| {
-        o.backup = true;
-        o.no_data = true;
-        o.backup_path = Some(custom_dir.clone());
-    });
-    assert_eq!(code, 0, "filter-repo-rs run should succeed");
-
-    let backup_dir = repo.join(&custom_dir);
-    assert!(
-        backup_dir.exists(),
-        "backup directory should exist at {:?}",
-        backup_dir
-    );
-    let bundles: Vec<_> = fs::read_dir(&backup_dir)
-        .expect("failed to read custom backup directory")
-        .filter_map(|entry| {
-            let entry = entry.ok()?;
-            let path = entry.path();
-            if path.extension().and_then(|ext| ext.to_str()) == Some("bundle") {
-                Some(path)
-            } else {
-                None
-            }
-        })
-        .collect();
-    assert!(
-        !bundles.is_empty(),
-        "expected at least one bundle in {:?}, entries: {:?}",
-        backup_dir,
-        bundles
-    );
-}
-
-#[test]
-fn backup_honors_explicit_file_path() {
-    let repo = init_repo();
-    let rel_path = PathBuf::from("custom/custom-bundle.bundle");
-    let expected_path = repo.join(&rel_path);
-    let (code, _o, _e) = run_tool(&repo, |o| {
-        o.backup = true;
-        o.no_data = true;
-        o.backup_path = Some(rel_path.clone());
-    });
-    assert_eq!(code, 0, "filter-repo-rs run should succeed");
-
-    assert!(
-        expected_path.exists(),
-        "expected bundle to exist at {:?}",
-        expected_path
-    );
-}
-
-#[test]
-fn branch_rename_updates_ref_and_head() {
-    let repo = init_repo();
-    // Determine current HEAD ref name
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    let headref = headref.trim().to_string(); // e.g., refs/heads/master
-                                              // Perform branch rename: prefix '' -> 'renamed-'
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((Vec::new(), b"renamed-".to_vec()));
-        o.no_data = true;
-    });
-    // Verify new branch exists and old deleted
-    let orig_name = headref.strip_prefix("refs/heads/").unwrap_or(&headref);
-    let new_branch = format!("refs/heads/renamed-{}", orig_name);
-    let (_c1, out1, _e1) = run_git(&repo, &["show-ref", "--verify", &new_branch]);
-    assert!(
-        !out1.is_empty(),
-        "expected new branch to exist: {}",
-        new_branch
-    );
-    let (_c2, out2, _e2) = run_git(&repo, &["show-ref", "--verify", &headref]);
-    assert!(
-        out2.is_empty(),
-        "expected old branch to be deleted: {}",
-        headref
-    );
-    // Verify HEAD points to new branch
-    let (_c3, head_after, _e3) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(
-        head_after.trim(),
-        new_branch,
-        "expected HEAD to follow renamed branch"
-    );
-}
-
-#[test]
-fn branch_rename_without_new_commits_updates_refs() {
-    let repo = init_repo();
-    // Create a branch pointing to existing commit and move HEAD there
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "feature/plain"]).0,
-        0
-    );
-    let (_c_before, head_before, _e_before) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(head_before.trim(), "refs/heads/feature/plain");
-
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((b"feature/".to_vec(), b"topic/".to_vec()));
-        o.no_data = true;
-    });
-
-    let (_c_new, out_new, _e_new) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/topic/plain"]);
-    assert!(
-        !out_new.is_empty(),
-        "expected refs/heads/topic/plain to exist"
-    );
-    let (_c_old, out_old, _e_old) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/feature/plain"]);
-    assert!(
-        out_old.is_empty(),
-        "expected refs/heads/feature/plain to be deleted"
-    );
-
-    let (_c_head, head_after, _e_head) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(
-        head_after.trim(),
-        "refs/heads/topic/plain",
-        "expected HEAD to follow renamed branch with no new commits"
-    );
-}
-
-#[test]
-fn branch_prefix_rename_preserves_head_to_mapped_target() {
-    let repo = init_repo();
-    // Create and switch to a prefixed branch
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "features/foo"]).0,
-        0
-    );
-    write_file(&repo, "feat.txt", "feat");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "feat commit"]).0, 0);
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(headref.trim(), "refs/heads/features/foo");
-    // Rename prefix features/ -> topics/
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((b"features/".to_vec(), b"topics/".to_vec()));
-        o.no_data = true;
-    });
-    // New branch exists, old deleted
-    let (_c1, out1, _e1) = run_git(&repo, &["show-ref", "--verify", "refs/heads/topics/foo"]);
-    assert!(!out1.is_empty(), "expected refs/heads/topics/foo to exist");
-    let (_c2, out2, _e2) = run_git(&repo, &["show-ref", "--verify", "refs/heads/features/foo"]);
-    assert!(
-        out2.is_empty(),
-        "expected refs/heads/features/foo to be deleted"
-    );
-    // HEAD moved to mapped target
-    let (_c3, head_after, _e3) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(
-        head_after.trim(),
-        "refs/heads/topics/foo",
-        "expected HEAD to follow mapped branch"
-    );
-}
-
-#[test]
-fn head_preserved_when_branch_unchanged() {
-    let repo = init_repo();
-    // Create another branch but keep HEAD on default branch
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    let headref = headref.trim().to_string();
-    assert!(headref.starts_with("refs/heads/"));
-    assert_eq!(run_git(&repo, &["checkout", "-q", "-b", "feature/x"]).0, 0);
-    // Switch back to original HEAD branch
-    assert_eq!(
-        run_git(
-            &repo,
-            &[
-                "checkout",
-                "-q",
-                headref.strip_prefix("refs/heads/").unwrap_or(&headref)
-            ]
-        )
-        .0,
-        0
-    );
-    // Rename a different prefix that doesn't match current HEAD branch
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((b"feature/".to_vec(), b"topic/".to_vec()));
-        o.no_data = true;
-    });
-    // HEAD should remain the same
-    let (_c1, head_after, _e1) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(head_after.trim(), headref, "expected HEAD to be preserved");
-}
-
-#[test]
-fn multi_branch_prefix_rename_maps_all_and_preserves_others() {
-    let repo = init_repo();
-    // Determine default branch name
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    let headref = headref.trim().to_string();
-    let def_short = headref
-        .strip_prefix("refs/heads/")
-        .unwrap_or(&headref)
-        .to_string();
-
-    // Create branches features/foo and features/bar with commits
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "features/foo"]).0,
-        0
-    );
-    write_file(&repo, "f-foo.txt", "foo");
-    run_git(&repo, &["add", "."]).0;
-    run_git(&repo, &["commit", "-q", "-m", "feat foo"]).0;
-    assert_eq!(run_git(&repo, &["checkout", "-q", &def_short]).0, 0);
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "features/bar"]).0,
-        0
-    );
-    write_file(&repo, "f-bar.txt", "bar");
-    run_git(&repo, &["add", "."]).0;
-    run_git(&repo, &["commit", "-q", "-m", "feat bar"]).0;
-
-    // Another branch that should remain unchanged
-    assert_eq!(run_git(&repo, &["checkout", "-q", &def_short]).0, 0);
-    assert_eq!(run_git(&repo, &["checkout", "-q", "-b", "misc/baz"]).0, 0);
-    write_file(&repo, "baz.txt", "baz");
-    run_git(&repo, &["add", "."]).0;
-    run_git(&repo, &["commit", "-q", "-m", "misc baz"]).0;
-
-    // Apply branch prefix rename: features/ -> topics/
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((b"features/".to_vec(), b"topics/".to_vec()));
-        o.no_data = true;
-    });
-
-    // Both features/* moved to topics/*
-    let (_c1, out_topics_foo, _e1) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/topics/foo"]);
-    assert!(
-        !out_topics_foo.is_empty(),
-        "expected refs/heads/topics/foo to exist"
-    );
-    let (_c2, out_topics_bar, _e2) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/topics/bar"]);
-    assert!(
-        !out_topics_bar.is_empty(),
-        "expected refs/heads/topics/bar to exist"
-    );
-
-    // Old branches deleted
-    let (_c3, out_features_foo, _e3) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/features/foo"]);
-    assert!(
-        out_features_foo.is_empty(),
-        "expected refs/heads/features/foo to be deleted"
-    );
-    let (_c4, out_features_bar, _e4) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/features/bar"]);
-    assert!(
-        out_features_bar.is_empty(),
-        "expected refs/heads/features/bar to be deleted"
-    );
-
-    // Unrelated branch intact
-    let (_c5, out_misc_baz, _e5) = run_git(&repo, &["show-ref", "--verify", "refs/heads/misc/baz"]);
-    assert!(
-        !out_misc_baz.is_empty(),
-        "expected refs/heads/misc/baz to remain"
-    );
-}
-
-#[test]
-fn multi_branch_prefix_rename_maps_head_from_deleted_branch() {
-    let repo = init_repo();
-    // Determine default branch name
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    let headref = headref.trim().to_string();
-    let def_short = headref
-        .strip_prefix("refs/heads/")
-        .unwrap_or(&headref)
-        .to_string();
-
-    // Create features/foo and features/bar, move HEAD to features/bar
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "features/foo"]).0,
-        0
-    );
-    write_file(&repo, "f-foo.txt", "foo");
-    run_git(&repo, &["add", "."]).0;
-    run_git(&repo, &["commit", "-q", "-m", "feat foo"]).0;
-    assert_eq!(run_git(&repo, &["checkout", "-q", &def_short]).0, 0);
-    assert_eq!(
-        run_git(&repo, &["checkout", "-q", "-b", "features/bar"]).0,
-        0
-    );
-    write_file(&repo, "f-bar.txt", "bar");
-    run_git(&repo, &["add", "."]).0;
-    run_git(&repo, &["commit", "-q", "-m", "feat bar"]).0;
-
-    // Verify HEAD is features/bar
-    let (_c_h, head_before, _e_h) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(head_before.trim(), "refs/heads/features/bar");
-
-    // Apply rename features/ -> topics/
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.branch_rename = Some((b"features/".to_vec(), b"topics/".to_vec()));
-        o.no_data = true;
-    });
-
-    // HEAD should map to topics/bar
-    let (_c1, head_after, _e1) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(
-        head_after.trim(),
-        "refs/heads/topics/bar",
-        "expected HEAD to map to renamed branch"
-    );
-}
-
-#[test]
-fn max_blob_size_drops_large_blobs() {
-    let repo = init_repo();
-    // Create files of different sizes
-    let big = vec![b'A'; 4096];
-    let small = vec![b'B'; 10];
-    fs::write(repo.join("big.bin"), &big).unwrap();
-    fs::write(repo.join("small.bin"), &small).unwrap();
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add blobs"]).0, 0);
-    // Run with max blob size threshold smaller than big.bin
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1024);
-        o.no_data = false;
-    });
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree.contains("small.bin"),
-        "expected small.bin to remain, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("big.bin"),
-        "expected big.bin to be dropped, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn replace_text_redacts_blob_contents() {
-    let repo = init_repo();
-    // Create a file with a secret token
-    write_file(&repo, "secret.txt", "token=SECRET-ABC-123\n");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add secret"]).0, 0);
-    // Replacement file
-    let repl = repo.join("repl-blobs.txt");
-    fs::write(&repl, "SECRET-ABC-123==>REDACTED\n").unwrap();
-    // Run tool with --replace-text
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.replace_text_file = Some(repl.clone());
-        o.no_data = false;
-    });
-    // Read back blob content via git show HEAD:secret.txt
-    let (_c2, content, _e2) = run_git(&repo, &["show", "HEAD:secret.txt"]);
-    assert!(
-        content.contains("REDACTED"),
-        "expected blob content to be redacted, got: {}",
-        content
-    );
-    assert!(
-        !content.contains("SECRET-ABC-123"),
-        "expected original secret to be removed, got: {}",
-        content
-    );
-}
-
-#[test]
-fn path_filter_includes_only_prefix() {
-    let repo = init_repo();
-    write_file(&repo, "src/keep.txt", "k");
-    write_file(&repo, "docs/drop.txt", "d");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add files"]).0, 0);
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.paths.push(b"src/".to_vec());
-    });
-    // Read back with quoting disabled for human-readable output
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        tree.contains("src/keep.txt"),
-        "expected to keep src/keep.txt, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("docs/drop.txt"),
-        "expected to drop docs/drop.txt, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn path_rename_applies_to_paths() {
-    let repo = init_repo();
-    write_file(&repo, "a/file.txt", "x");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add a/file.txt"]).0,
-        0
-    );
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.path_renames.push((b"a/".to_vec(), b"x/".to_vec()));
-    });
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree.contains("x/file.txt"),
-        "expected path renamed to x/file.txt, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("a/file.txt"),
-        "expected old path a/file.txt removed, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn path_glob_selects_md_under_src() {
-    let repo = init_repo();
-    write_file(&repo, "src/a.md", "m");
-    write_file(&repo, "src/a.txt", "t");
-    write_file(&repo, "src/deep/b.md", "m");
-    write_file(&repo, "docs/x.md", "m");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add various files"]).0,
-        0
-    );
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.path_globs.push(b"src/**/*.md".to_vec());
-    });
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree.contains("src/a.md"),
-        "expected to keep src/a.md, got: {}",
-        tree
-    );
-    assert!(
-        tree.contains("src/deep/b.md"),
-        "expected to keep src/deep/b.md, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("src/a.txt"),
-        "expected to drop src/a.txt, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("docs/x.md"),
-        "expected to drop docs/x.md, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn path_regex_filters_and_respects_invert() {
-    // Direct regex include (keep *.rs only)
-    let repo = init_repo();
-    write_file(&repo, "src/lib.rs", "fn main() {}
-");
-    write_file(&repo, "README.md", "docs
-");
-    write_file(&repo, "scripts/build.sh", "echo hi
-");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "seed files"]).0,
-        0
-    );
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.path_regexes.push(Regex::new(r".*\.rs$").unwrap());
-    });
-    let (_c_tree, tree, _e_tree) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(tree.contains("src/lib.rs"), "expected regex to keep src/lib.rs: {}", tree);
-    assert!(
-        !tree.contains("README.md"),
-        "expected regex to drop README.md: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("scripts/build.sh"),
-        "expected regex to drop scripts/build.sh: {}",
-        tree
-    );
-
-    // Inverted regex (drop *.md, keep others)
-    let repo2 = init_repo();
-    write_file(&repo2, "src/main.rs", "fn main() {}
-");
-    write_file(&repo2, "docs/readme.md", "docs
-");
-    write_file(&repo2, "notes/todo.txt", "todo
-");
-    run_git(&repo2, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo2, &["commit", "-q", "-m", "seed docs"]).0,
-        0
-    );
-    let (_c2, _o2, _e2) = run_tool(&repo2, |o| {
-        o.path_regexes.push(Regex::new(r".*\.md$").unwrap());
-        o.invert_paths = true;
-    });
-    let (_c_tree2, tree2, _e_tree2) = run_git(
-        &repo2,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        tree2.contains("src/main.rs"),
-        "expected inverted regex to keep src/main.rs: {}",
-        tree2
-    );
-    assert!(
-        tree2.contains("notes/todo.txt"),
-        "expected inverted regex to keep notes/todo.txt: {}",
-        tree2
-    );
-    assert!(
-        !tree2.contains("docs/readme.md"),
-        "expected inverted regex to drop docs/readme.md: {}",
-        tree2
-    );
-}
-
-#[test]
-fn path_filter_and_rename_updates_commit_and_ref_maps() {
-    let repo = init_repo();
-
-    // Create a branch that will be renamed and populate src/ with content to keep.
-    assert_eq!(run_git(&repo, &["checkout", "-b", "feature/topic"]).0, 0);
-    write_file(&repo, "src/lib.rs", "fn main() { println!(\"hi\"); }\n");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add src content"]).0,
-        0
-    );
-    assert_eq!(
-        run_git(&repo, &["tag", "-a", "-m", "cut first release", "v1"]).0,
-        0
-    );
-
-    let (_c_old, old_head, _e_old) = run_git(&repo, &["rev-parse", "HEAD"]);
-    let old_head = old_head.trim().to_string();
-
-    let (code, _o, _e) = run_tool(&repo, |o| {
-        o.paths.push(b"src/".to_vec());
-        o.path_renames
-            .push((b"src/".to_vec(), b"app/".to_vec()));
-        o.branch_rename = Some((b"feature/".to_vec(), b"topics/".to_vec()));
-        o.tag_rename = Some((b"v".to_vec(), b"release-".to_vec()));
-    });
-    assert_eq!(code, 0, "filter-repo-rs run should succeed");
-
-    let (_c_new, new_head, _e_new) = run_git(&repo, &["rev-parse", "HEAD"]);
-    let new_head = new_head.trim().to_string();
-
-    // Only the renamed path should remain in the tree.
-    let (_c_tree, tree, _e_tree) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(tree.contains("app/lib.rs"), "expected renamed path in tree: {}", tree);
-    assert!(
-        !tree.contains("src/lib.rs"),
-        "expected original path filtered out: {}",
-        tree
-    );
-
-    // commit-map should capture the old -> new HEAD mapping.
-    let commit_map = repo
-        .join(".git")
-        .join("filter-repo")
-        .join("commit-map");
-    let cm = std::fs::read_to_string(&commit_map).unwrap();
-    assert!(
-        cm.contains(&format!("{} {}", old_head, new_head)),
-        "commit-map missing HEAD mapping; contents: {}",
-        cm
-    );
-
-    // ref-map should record branch and tag renames.
-    let ref_map = repo.join(".git").join("filter-repo").join("ref-map");
-    let rm = std::fs::read_to_string(&ref_map).unwrap();
-    assert!(
-        rm.contains("refs/heads/feature/topic refs/heads/topics/topic"),
-        "ref-map missing branch rename: {}",
-        rm
-    );
-    assert!(
-        rm.contains("refs/tags/v1 refs/tags/release-1"),
-        "ref-map missing tag rename: {}",
-        rm
-    );
-
-    // Verify final refs reflect the ref-map entries.
-    let (_c_branch, branch_out, _e_branch) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/topics/topic"]);
-    assert!(
-        !branch_out.is_empty(),
-        "expected renamed branch to exist: {}",
-        branch_out
-    );
-    let (_c_branch_old, branch_old_out, _e_branch_old) =
-        run_git(&repo, &["show-ref", "--verify", "refs/heads/feature/topic"]);
-    assert!(
-        branch_old_out.is_empty(),
-        "expected original branch to be removed: {}",
-        branch_old_out
-    );
-
-    let (_c_tag, tag_out, _e_tag) =
-        run_git(&repo, &["show-ref", "--verify", "refs/tags/release-1"]);
-    assert!(
-        !tag_out.is_empty(),
-        "expected renamed tag to exist: {}",
-        tag_out
-    );
-    let (_c_tag_old, tag_old_out, _e_tag_old) =
-        run_git(&repo, &["show-ref", "--verify", "refs/tags/v1"]);
-    assert!(
-        tag_old_out.is_empty(),
-        "expected original tag to be removed: {}",
-        tag_old_out
-    );
-
-    // HEAD should now point at the renamed branch.
-    let (_c_head, head_ref, _e_head) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    assert_eq!(
-        head_ref.trim(),
-        "refs/heads/topics/topic",
-        "HEAD should track renamed branch"
-    );
-}
-
-#[test]
-fn windows_path_policy_sanitizes_or_preserves_bytes() {
-    let cases = vec![
-        (b"dir/inv:name?.txt ".as_ref(), b"dir/inv_name_.txt".as_ref()),
-        (b"dir/trailing.dot.".as_ref(), b"dir/trailing.dot".as_ref()),
-        (b"simple.txt".as_ref(), b"simple.txt".as_ref()),
-    ];
-
-    for (input, expected_windows) in cases {
-        let sanitized = fr::pathutil::sanitize_invalid_windows_path_bytes(input);
-        if cfg!(windows) {
-            assert_eq!(
-                sanitized,
-                expected_windows.to_vec(),
-                "windows sanitization mismatch for {:?}",
-                String::from_utf8_lossy(input)
-            );
-        } else {
-            assert_eq!(
-                sanitized,
-                input.to_vec(),
-                "non-windows platforms should preserve bytes"
-            );
-        }
-    }
-}
-
-#[test]
-fn invert_paths_drops_prefix() {
-    let repo = init_repo();
-    write_file(&repo, "src/keep.txt", "k");
-    write_file(&repo, "drop/file.txt", "d");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "prepare files"]).0,
-        0
-    );
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.paths.push(b"drop/".to_vec());
-        o.invert_paths = true;
-    });
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree.contains("src/keep.txt"),
-        "expected to keep src/keep.txt, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("drop/file.txt"),
-        "expected to drop drop/file.txt, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn quoted_paths_roundtrip_with_rename() {
-    let repo = init_repo();
-    // Force quoting in fast-export
-    run_git(&repo, &["config", "core.quotepath", "true"]);
-    // Create a filename with non-ASCII to trigger C-style quoting
-    write_file(&repo, "src/ümlaut.txt", "u");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add umlaut"]).0, 0);
-    // Move everything into X/ using to-subdirectory behavior; disable our quotepath override
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.quotepath = false; // do NOT pass -c core.quotepath=false
-        o.path_renames.push((Vec::new(), b"X/".to_vec()));
-        o.no_data = true;
-    });
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    // Normalize by dequoting any C-style quoted entries
-    let mut found = false;
-    for line in tree.lines() {
-        let s = line.trim();
-        if s.is_empty() {
-            continue;
-        }
-        let norm = if s.starts_with('"') && s.ends_with('"') && s.len() >= 2 {
-            let inner = &s.as_bytes()[1..s.as_bytes().len() - 1];
-            String::from_utf8_lossy(&fr::dequote_c_style_bytes(inner)).to_string()
-        } else {
-            s.to_string()
-        };
-        if norm == "X/src/ümlaut.txt" {
-            found = true;
-            break;
-        }
-    }
-    assert!(
-        found,
-        "expected renamed quoted path to roundtrip, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn rename_and_copy_paths_requote_after_filtering() {
-    let repo = init_repo();
-    let stream_path = repo.join("fe-renames.stream");
-    let stream = r#"blob
-mark :1
-data 4
-one
-
-commit refs/heads/main
-mark :2
-author Tester <tester@example.com> 0 +0000
-committer Tester <tester@example.com> 0 +0000
-data 3
-c1
-M 100644 :1 "sp ace.txt"
-M 100644 :1 "old\001.txt"
-M 100644 :1 "removed space.txt"
-
-commit refs/heads/main
-mark :3
-author Tester <tester@example.com> 1 +0000
-committer Tester <tester@example.com> 1 +0000
-data 3
-c2
-from :2
-D "removed space.txt"
-C "sp ace.txt" "dup space.txt"
-R "old\001.txt" "final\001name.txt"
-
-done
-"#;
-    fs::write(&stream_path, stream).expect("write custom fast-export stream");
-
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.dry_run = true;
-        o.path_renames.push((Vec::new(), b"prefix/".to_vec()));
-        #[allow(deprecated)]
-        {
-            o.fe_stream_override = Some(stream_path.clone());
-        }
-    });
-
-    let filtered_path = repo
-        .join(".git")
-        .join("filter-repo")
-        .join("fast-export.filtered");
-    let filtered = fs::read_to_string(&filtered_path).expect("read filtered stream");
-
-    assert!(
-        filtered.contains("M 100644 :1 \"prefix/sp ace.txt\""),
-        "expected prefixed modify line, got: {}",
-        filtered
-    );
-    assert!(
-        filtered.contains("M 100644 :1 \"prefix/old\\001.txt\""),
-        "expected prefixed control-char modify line, got: {}",
-        filtered
-    );
-    assert!(
-        filtered.contains("D \"prefix/removed space.txt\""),
-        "expected prefixed delete line, got: {}",
-        filtered
-    );
-    assert!(
-        filtered.contains("C \"prefix/sp ace.txt\" \"prefix/dup space.txt\""),
-        "expected prefixed copy line, got: {}",
-        filtered
-    );
-    assert!(
-        filtered.contains("R \"prefix/old\\001.txt\" \"prefix/final\\001name.txt\""),
-        "expected prefixed rename line, got: {}",
-        filtered
-    );
-}
-
-#[test]
-fn inline_replace_text_and_report_modified() {
-    // use std::env as stdenv;
-    let repo = init_repo();
-    // Build a minimal fast-export-like stream that uses inline data in a commit
-    let stream_path = repo.join("fe-inline.stream");
-    let payload = "token=SECRET-INLINE-123\n";
-    let payload_len = payload.as_bytes().len();
-    let msg = "inline commit\n";
-    let msg_len = msg.as_bytes().len();
-    let mut s = String::new();
-    let (_hc, headref, _he) = run_git(&repo, &["symbolic-ref", "-q", "HEAD"]);
-    let commit_ref = headref.trim();
-    s.push_str(&format!("commit {}\n", commit_ref));
-    s.push_str("mark :1\n");
-    s.push_str("committer A U Thor <a.u.thor@example.com> 1737070000 +0000\n");
-    s.push_str(&format!("data {}\n{}", msg_len, msg));
-    s.push_str("M 100644 inline secret.txt\n");
-    s.push_str(&format!("data {}\n{}", payload_len, payload));
-    s.push_str("\n");
-    s.push_str("done\n");
-    std::fs::write(&stream_path, s).unwrap();
-
-    // Replacement rules: redact the secret token
-    let repl = repo.join("repl-inline.txt");
-    std::fs::write(&repl, "SECRET-INLINE-123==>REDACTED\n").unwrap();
-
-    // Force our tool to read the prebuilt stream via options override
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.replace_text_file = Some(repl.clone());
-        o.no_data = false;
-        o.write_report = true;
-        // use internal test override
-        #[allow(deprecated)]
-        {
-            o.fe_stream_override = Some(stream_path.clone());
-        }
-    });
-
-    // Verify blob content rewritten
-    let (_cc, content, _ee) = run_git(&repo, &["show", "HEAD:secret.txt"]);
-    assert!(
-        content.contains("REDACTED"),
-        "expected inline blob to be redacted, got: {}",
-        content
-    );
-    assert!(
-        !content.contains("SECRET-INLINE-123"),
-        "expected original secret to be removed, got: {}",
-        content
-    );
-
-    // Verify report includes modified count and path sample
-    let report = repo.join(".git").join("filter-repo").join("report.txt");
-    let mut s = String::new();
-    std::fs::File::open(&report)
-        .unwrap()
-        .read_to_string(&mut s)
-        .unwrap();
-    assert!(
-        s.contains("Blobs modified by replace-text"),
-        "expected modified counter in report, got: {}",
-        s
-    );
-    assert!(
-        s.contains("secret.txt"),
-        "expected modified sample path in report, got: {}",
-        s
-    );
-}
-#[test]
-fn replace_text_regex_redacts_blob() {
-    let repo = init_repo();
-    write_file(&repo, "data.txt", "foo123 foo999\n");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add data"]).0, 0);
-    let repl = repo.join("repl-regex.txt");
-    std::fs::write(&repl, "regex:foo[0-9]+==>X\n").unwrap();
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.replace_text_file = Some(repl.clone());
-        o.no_data = false;
-    });
-    let (_c2, content, _e2) = run_git(&repo, &["show", "HEAD:data.txt"]);
-    assert!(
-        content.contains("X X"),
-        "expected both occurrences replaced, got: {}",
-        content
-    );
-    assert!(
-        !content.contains("foo123"),
-        "expected original tokens removed, got: {}",
-        content
-    );
-}
-
-#[test]
-fn strip_report_written() {
-    let repo = init_repo();
-    // create one small and one large file
-    write_file(&repo, "small.txt", "x");
-    // ~10KB large file
-    let big_data = vec![b'A'; 10_000];
-    let mut f = File::create(repo.join("big.bin")).unwrap();
-    f.write_all(&big_data).unwrap();
-    f.flush().unwrap();
-    drop(f);
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add files"]).0, 0);
-
-    // First verify that big.bin exists before filtering
-    let (_c1, tree_before, _e1) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree_before.contains("big.bin"),
-        "big.bin should exist before filtering: {}",
-        tree_before
-    );
-
-    // run tool with small max-blob-size and report enabled
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1024);
-        o.write_report = true;
-    });
-
-    // Verify that big.bin was actually filtered out
-    let (_c2, tree_after, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        !tree_after.contains("big.bin"),
-        "big.bin should be filtered out: {}",
-        tree_after
-    );
-    assert!(
-        tree_after.contains("small.txt"),
-        "small.txt should remain: {}",
-        tree_after
-    );
-
-    let report = repo.join(".git").join("filter-repo").join("report.txt");
-    assert!(report.exists(), "expected report at {:?}", report);
-    let mut s = String::new();
-    File::open(&report).unwrap().read_to_string(&mut s).unwrap();
-
-    // The report should indicate that blobs were stripped by size
-    assert!(
-        s.contains("Blobs stripped by size"),
-        "expected size counter in report, got: {}",
-        s
-    );
-
-    // Either the count should be > 0 OR the sample paths should contain big.bin
-    let has_count =
-        s.contains("Blobs stripped by size: 1") || s.contains("Blobs stripped by size: 2");
-    let has_sample = s.contains("big.bin")
-        || s.contains("Sample paths (size):") && s.lines().any(|l| l.trim() == "big.bin");
-
-    assert!(
-        has_count || has_sample,
-        "Expected either count > 0 or big.bin sample in report, got: {}",
-        s
-    );
-}
-
-#[test]
-fn dry_run_does_not_modify_refs_or_remote() {
-    let repo = init_repo();
-    // Track HEAD and add a self origin
-    let (_c0, head_before, _e0) = run_git(&repo, &["rev-parse", "HEAD"]);
-    assert_eq!(run_git(&repo, &["remote", "add", "origin", "."]).0, 0);
-    // Run with dry-run and write-report
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.dry_run = true;
-        o.write_report = true;
-        o.no_data = true;
-    });
-    // HEAD unchanged
-    let (_c1, head_after, _e1) = run_git(&repo, &["rev-parse", "HEAD"]);
-    assert_eq!(
-        head_before.trim(),
-        head_after.trim(),
-        "expected HEAD unchanged in dry-run"
-    );
-    // origin remote still exists
-    let (_c2, remotes, _e2) = run_git(&repo, &["remote"]);
-    assert!(
-        remotes.contains("origin"),
-        "expected origin to remain in dry-run, got: {}",
-        remotes
-    );
-    // report exists
-    let report = repo.join(".git").join("filter-repo").join("report.txt");
-    assert!(
-        report.exists(),
-        "expected report in dry-run at {:?}",
-        report
-    );
-}
-
-#[test]
-fn strip_ids_report_written() {
-    let repo = init_repo();
-    // Create a file and commit
-    write_file(&repo, "secret.bin", "topsecret\n");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add secret.bin"]).0,
-        0
-    );
-    // Resolve blob id for HEAD:secret.bin
-    let (_c0, blob_id, _e0) = run_git(&repo, &["rev-parse", "HEAD:secret.bin"]);
-    let sha = blob_id.trim();
-    // Write SHA to list file
-    let shalist = repo.join("strip-sha.txt");
-    std::fs::write(&shalist, format!("{}\n", sha)).unwrap();
-    // Run tool with strip-blobs-with-ids and report enabled
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.strip_blobs_with_ids = Some(shalist.clone());
-        o.write_report = true;
-    });
-    // The file should be dropped from the tree
-    let (_c1, tree, _e1) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        !tree.contains("secret.bin"),
-        "expected secret.bin to be dropped, got: {}",
-        tree
-    );
-    // Report should include SHA count and sample path under sha section
-    let report = repo.join(".git").join("filter-repo").join("report.txt");
-    let mut s = String::new();
-    std::fs::File::open(&report)
-        .unwrap()
-        .read_to_string(&mut s)
-        .unwrap();
-    assert!(
-        s.contains("Blobs stripped by SHA:"),
-        "expected SHA counter in report, got: {}",
-        s
-    );
-    assert!(
-        s.contains("Sample paths (sha):") && s.contains("secret.bin"),
-        "expected sha sample path in report, got: {}",
-        s
-    );
-}
-
-#[test]
-fn partial_mode_keeps_origin_and_remote_tracking() {
-    let repo = init_repo();
-    // Determine current HEAD ref
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "-q", "HEAD"]);
-    let headref = headref.trim();
-    let branch = headref.strip_prefix("refs/heads/").unwrap_or(headref);
-    // Add a self origin and create a remote-tracking ref
-    assert_eq!(run_git(&repo, &["remote", "add", "origin", "."]).0, 0);
-    let spec = format!("+{}:refs/remotes/origin/{}", headref, branch);
-    assert_eq!(run_git(&repo, &["fetch", "origin", &spec]).0, 0);
-    // Run tool with partial
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-        o.partial = true;
-    });
-    // origin remote should remain
-    let (_c2, remotes, _e2) = run_git(&repo, &["remote"]);
-    assert!(
-        remotes.contains("origin"),
-        "expected origin remote to remain in partial mode, got: {}",
-        remotes
-    );
-    // remote-tracking ref should remain
-    let (c3, _o3, _e3) = run_git(
-        &repo,
-        &[
-            "show-ref",
-            "--verify",
-            &format!("refs/remotes/origin/{}", branch),
-        ],
-    );
-    assert_eq!(
-        c3, 0,
-        "expected remote-tracking ref to remain in partial mode"
-    );
-}
-
-#[test]
-fn sensitive_fetch_all_from_bare_remote() {
-    // Create a bare remote with an extra branch not present locally
-    let bare = mktemp("fr_rs_bare");
-    std::fs::create_dir_all(&bare).unwrap();
-    assert_eq!(run_git(&bare, &["init", "--bare"]).0, 0);
-
-    // Seed repo to push into bare
-    let seed = mktemp("fr_rs_seed");
-    std::fs::create_dir_all(&seed).unwrap();
-    assert_eq!(run_git(&seed, &["init"]).0, 0);
-    run_git(&seed, &["config", "user.name", "A U Thor"]).0;
-    run_git(&seed, &["config", "user.email", "a.u.thor@example.com"]).0;
-    write_file(&seed, "README.md", "seed");
-    run_git(&seed, &["add", "."]).0;
-    assert_eq!(run_git(&seed, &["commit", "-q", "-m", "seed init"]).0, 0);
-    // Create an extra branch with a file
-    assert_eq!(run_git(&seed, &["checkout", "-b", "extra"]).0, 0);
-    write_file(&seed, "extra.txt", "x");
-    run_git(&seed, &["add", "."]).0;
-    assert_eq!(run_git(&seed, &["commit", "-q", "-m", "add extra"]).0, 0);
-    // Add bare as origin and push all
-    let bare_str = bare.to_string_lossy().to_string();
-    assert_eq!(run_git(&seed, &["remote", "add", "origin", &bare_str]).0, 0);
-    assert_eq!(run_git(&seed, &["push", "-q", "origin", "--all"]).0, 0);
-
-    // Our working repo
-    let repo = init_repo();
-    // Add origin pointing to bare, but do not fetch
-    assert_eq!(run_git(&repo, &["remote", "add", "origin", &bare_str]).0, 0);
-
-    // Sanity: branch 'extra' should not exist locally yet
-    let (c0, _o0, _e0) = run_git(&repo, &["show-ref", "--verify", "refs/heads/extra"]);
-    assert_ne!(
-        c0, 0,
-        "extra branch should not exist before sensitive fetch"
-    );
-
-    // Run tool in sensitive mode to trigger fetch-all (no --no-fetch)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.sensitive = true;
-        o.no_data = true;
-    });
-
-    // After run, 'extra' should exist due to fetch-all with empty refmap
-    let (c1, _o1, _e1) = run_git(&repo, &["show-ref", "--verify", "refs/heads/extra"]);
-    assert_eq!(
-        c1, 0,
-        "expected sensitive fetch-all to create refs/heads/extra"
-    );
-    // origin remote should remain in sensitive mode
-    let (_c2, remotes, _e2) = run_git(&repo, &["remote"]);
-    assert!(
-        remotes.contains("origin"),
-        "expected origin to remain for sensitive mode"
-    );
-}
-#[test]
-fn origin_migration_and_removal_nonsensitive() {
-    let repo = init_repo();
-    // Determine current HEAD ref
-    let (_c0, headref, _e0) = run_git(&repo, &["symbolic-ref", "-q", "HEAD"]);
-    let headref = headref.trim();
-    let branch = headref.strip_prefix("refs/heads/").unwrap_or(headref);
-    // Add a self origin and create a remote-tracking ref
-    assert_eq!(run_git(&repo, &["remote", "add", "origin", "."]).0, 0);
-    let spec = format!("+{}:refs/remotes/origin/{}", headref, branch);
-    assert_eq!(run_git(&repo, &["fetch", "origin", &spec]).0, 0);
-    // Verify remote-tracking exists
-    let (_c1, out1, _e1) = run_git(
-        &repo,
-        &[
-            "show-ref",
-            "--verify",
-            &format!("refs/remotes/origin/{}", branch),
-        ],
-    );
-    assert!(
-        out1.contains(&format!("refs/remotes/origin/{}", branch)),
-        "expected remote-tracking ref created, got: {}",
-        out1
-    );
-    // Run tool (non-sensitive, full)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-    });
-    // Origin remote should be removed
-    let (_c2, remotes, _e2) = run_git(&repo, &["remote"]);
-    assert!(
-        !remotes.contains("origin"),
-        "expected origin remote removed, got: {}",
-        remotes
-    );
-    // Remote-tracking ref should be gone
-    let (c3, _o3, _e3) = run_git(
-        &repo,
-        &[
-            "show-ref",
-            "--verify",
-            &format!("refs/remotes/origin/{}", branch),
-        ],
-    );
-    assert_ne!(c3, 0, "expected remote-tracking ref removed");
-    // Local branch should exist
-    let (c4, _o4, _e4) = run_git(&repo, &["show-ref", "--verify", headref]);
-    assert_eq!(c4, 0, "expected local branch to exist: {}", headref);
-}
-
-#[test]
-fn sensitive_mode_keeps_origin_remote() {
-    let repo = init_repo();
-    // Ensure origin exists
-    assert_eq!(run_git(&repo, &["remote", "add", "origin", "."]).0, 0);
-    // Run tool with sensitive mode and no_fetch (avoid heavy operations)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.no_data = true;
-        o.sensitive = true;
-        o.no_fetch = true;
-    });
-    // Origin remote should remain
-    let (_c2, remotes, _e2) = run_git(&repo, &["remote"]);
-    assert!(
-        remotes.contains("origin"),
-        "expected origin remote to remain in sensitive mode, got: {}",
-        remotes
-    );
-}
-
-#[test]
-fn max_blob_size_edge_cases() {
-    let repo = init_repo();
-
-    // Test case 1: Empty file (size = 0)
-    write_file(&repo, "empty.txt", "");
-
-    // Test case 2: Single byte file (size = 1)
-    write_file(&repo, "tiny.txt", "A");
-
-    // Test case 3: Exactly at threshold
-    let threshold_content = vec![b'X'; 100];
-    fs::write(repo.join("threshold.bin"), &threshold_content).unwrap();
-
-    // Test case 4: Just over threshold
-    let over_content = vec![b'Y'; 101];
-    fs::write(repo.join("over.bin"), &over_content).unwrap();
-
-    // Test case 5: Very large file (stress test)
-    let large_content = vec![b'Z'; 10000];
-    fs::write(repo.join("large.bin"), &large_content).unwrap();
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add edge case files"]).0,
-        0
-    );
-
-    // Run with threshold = 100 - should keep files <= 100, drop files > 100
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(100);
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep empty.txt (0 bytes), tiny.txt (1 byte), threshold.bin (100 bytes)
-    assert!(
-        tree.contains("empty.txt"),
-        "expected empty.txt to remain, got: {}",
-        tree
-    );
-    assert!(
-        tree.contains("tiny.txt"),
-        "expected tiny.txt to remain, got: {}",
-        tree
-    );
-    assert!(
-        tree.contains("threshold.bin"),
-        "expected threshold.bin to remain, got: {}",
-        tree
-    );
-
-    // Should drop over.bin (101 bytes) and large.bin (10000 bytes)
-    assert!(
-        !tree.contains("over.bin"),
-        "expected over.bin to be dropped, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("large.bin"),
-        "expected large.bin to be dropped, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_with_path_filtering() {
-    let repo = init_repo();
-
-    // Create files in different directories
-    std::fs::create_dir_all(repo.join("keep")).unwrap();
-    std::fs::create_dir_all(repo.join("drop")).unwrap();
-    let large_content = vec![b'A'; 2000];
-    std::fs::write(repo.join("keep/large.bin"), &large_content).unwrap();
-    std::fs::write(repo.join("drop/large.bin"), &large_content).unwrap();
-    std::fs::write(repo.join("keep/small.txt"), "small content").unwrap();
-    std::fs::write(repo.join("drop/small.txt"), "small content").unwrap();
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(
-            &repo,
-            &["commit", "-q", "-m", "add files in different directories"]
-        )
-        .0,
-        0
-    );
-
-    // Filter to only keep/ directory AND drop blobs > 1000 bytes
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000);
-        o.paths.push(vec![b'k', b'e', b'e', b'p', b'/']); // "keep/"
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep keep/small.txt (within path filter and size limit)
-    assert!(
-        tree.contains("keep/small.txt"),
-        "expected keep/small.txt to remain, got: {}",
-        tree
-    );
-
-    // Should drop drop/ files (not in path filter)
-    assert!(
-        !tree.contains("drop/"),
-        "expected drop/ directory to be filtered out, got: {}",
-        tree
-    );
-
-    // Should drop keep/large.bin (in path filter but over size limit)
-    assert!(
-        !tree.contains("keep/large.bin"),
-        "expected keep/large.bin to be dropped due to size, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_with_strip_blobs_by_sha() {
-    let repo = init_repo();
-
-    // Create files with specific content to get predictable SHAs
-    let content1 = "test content 1";
-    let content2 = "test content 2";
-    fs::write(repo.join("file1.txt"), content1).unwrap();
-    fs::write(repo.join("file2.txt"), content2).unwrap();
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add test files"]).0,
-        0
-    );
-
-    // Get SHAs of the blobs
-    let (_c1, sha1_output, _e1) = run_git(&repo, &["hash-object", "file1.txt"]);
-    let (_c2, sha2_output, _e2) = run_git(&repo, &["hash-object", "file2.txt"]);
-    let sha1 = sha1_output.trim();
-    let sha2 = sha2_output.trim();
-
-    // Create SHA list file
-    let sha_list_content = format!("{}\n{}", sha1, sha2);
-    fs::write(repo.join("sha_list.txt"), &sha_list_content).unwrap();
-
-    run_git(&repo, &["add", "sha_list.txt"]).0;
-    assert_eq!(run_git(&repo, &["commit", "-q", "-m", "add sha list"]).0, 0);
-
-    // Run with both size filter and SHA filter
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000); // Should keep both files based on size
-        o.strip_blobs_with_ids = Some(repo.join("sha_list.txt"));
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Both files should be dropped due to SHA filter, regardless of size
-    assert!(
-        !tree.contains("file1.txt"),
-        "expected file1.txt to be dropped by SHA filter, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("file2.txt"),
-        "expected file2.txt to be dropped by SHA filter, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_empty_repository() {
-    let repo = init_repo();
-
-    // Run max-blob-size filter on repository with no blobs (just initial commit)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000);
-    });
-
-    // Should complete successfully without errors
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should still have README.md from initial repo
-    assert!(
-        tree.contains("README.md"),
-        "expected README.md to remain, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_mixed_blob_types() {
-    let repo = init_repo();
-
-    // Test different types of content
-    write_file(&repo, "text.txt", &"a".repeat(1500)); // Text content
-    fs::write(repo.join("binary.bin"), vec![0u8; 1500]).unwrap(); // Binary content
-    write_file(&repo, "utf8.txt", &"你好".repeat(500)); // UTF-8 content (500 chars * 3 bytes each = 1500 bytes)
-    fs::write(repo.join("zeroes.bin"), vec![0u8; 500]).unwrap(); // Small binary
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add mixed content types"]).0,
-        0
-    );
-
-    // Filter with threshold that should keep small files, drop large ones
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000);
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep small files
-    assert!(
-        tree.contains("zeroes.bin"),
-        "expected zeroes.bin to remain, got: {}",
-        tree
-    );
-
-    // Should drop large files regardless of content type
-    assert!(
-        !tree.contains("text.txt"),
-        "expected text.txt to be dropped due to size, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("binary.bin"),
-        "expected binary.bin to be dropped due to size, got: {}",
-        tree
-    );
-    assert!(
-        !tree.contains("utf8.txt"),
-        "expected utf8.txt to be dropped due to size, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_threshold_boundary() {
-    let repo = init_repo();
-
-    // Test exact boundary conditions
-    let exact_content = vec![b'X'; 1024]; // Exactly 1KB
-    let just_over = vec![b'Y'; 1025]; // Just over 1KB
-
-    fs::write(repo.join("exact.txt"), &exact_content).unwrap();
-    fs::write(repo.join("over.txt"), &just_over).unwrap();
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add boundary test files"]).0,
-        0
-    );
-
-    // Test with threshold = 1024
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1024);
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep exact.txt (exactly 1024 bytes)
-    assert!(
-        tree.contains("exact.txt"),
-        "expected exact.txt to remain, got: {}",
-        tree
-    );
-
-    // Should drop over.txt (1025 bytes)
-    assert!(
-        !tree.contains("over.txt"),
-        "expected over.txt to be dropped, got: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_batch_optimization_verification() {
-    let repo = init_repo();
-
-    // Create many files to make individual calls inefficient
-    for i in 0..100 {
-        let content = format!("file content {}", i);
-        write_file(&repo, &format!("file{}.txt", i), &content);
-    }
-
-    // Add a few large files that should be filtered
-    write_file(&repo, "large1.bin", &"a".repeat(2000));
-    write_file(&repo, "large2.bin", &"b".repeat(3000));
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(
-            &repo,
-            &["commit", "-q", "-m", "add many files for batch test"]
-        )
-        .0,
-        0
-    );
-
-    // Filter with size threshold - this should use batch optimization
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1500);
-    });
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep all small files
-    for i in 0..100 {
-        assert!(
-            tree.contains(&format!("file{}.txt", i)),
-            "expected file{}.txt to remain",
-            i
-        );
-    }
-
-    // Should drop large files
-    assert!(
-        !tree.contains("large1.bin"),
-        "expected large1.bin to be dropped due to size"
-    );
-    assert!(
-        !tree.contains("large2.bin"),
-        "expected large2.bin to be dropped due to size"
-    );
-
-    // Verify the total count is correct (100 small files should remain)
-    let files: Vec<&str> = tree.lines().collect();
-
-    // Debug: print all files to see what's unexpected
-    println!("Found {} files:", files.len());
-    for file in &files {
-        println!("  {}", file);
-    }
-
-    // Should be 100 small files, but account for any unexpected files
-    let small_files: Vec<&str> = files
-        .iter()
-        .copied()
-        .filter(|f| f.starts_with("file"))
-        .collect();
-    assert_eq!(
-        small_files.len(),
-        100,
-        "expected exactly 100 small files, got {}",
-        small_files.len()
-    );
-    assert!(
-        !tree.contains("large1.bin"),
-        "expected large1.bin to be dropped due to size"
-    );
-    assert!(
-        !tree.contains("large2.bin"),
-        "expected large2.bin to be dropped due to size"
-    );
-}
-
-#[test]
-fn max_blob_size_performance_comparison() {
-    let repo = init_repo();
-
-    // Create many files with varying sizes to test performance characteristics
-    for i in 0..50 {
-        write_file(&repo, &format!("small{}.txt", i), &"x".repeat(100)); // 100 bytes
-    }
-    for i in 0..20 {
-        write_file(&repo, &format!("medium{}.bin", i), &"y".repeat(1000)); // 1000 bytes
-    }
-    for i in 0..10 {
-        write_file(&repo, &format!("large{}.dat", i), &"z".repeat(5000)); // 5000 bytes
-    }
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add performance test files"]).0,
-        0
-    );
-
-    // Test that batch processing can handle the workload efficiently
-    let start = std::time::Instant::now();
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1500);
-    });
-    let duration = start.elapsed();
-
-    // Should complete in reasonable time (adjust threshold as needed)
-    assert!(
-        duration.as_secs() < 30,
-        "batch processing should complete quickly, took {:?}",
-        duration
-    );
-
-    let (_c2, tree, _e2) = run_git(&repo, &["ls-tree", "-r", "--name-only", "HEAD"]);
-
-    // Should keep small files (100 bytes)
-    for i in 0..50 {
-        assert!(
-            tree.contains(&format!("small{}.txt", i)),
-            "expected small{}.txt to remain",
-            i
-        );
-    }
-
-    // Should keep medium files (1000 bytes) since threshold is 1500
-    for i in 0..20 {
-        assert!(
-            tree.contains(&format!("medium{}.bin", i)),
-            "expected medium{}.bin to remain",
-            i
-        );
-    }
-
-    // Should drop large files (5000 bytes)
-    for i in 0..10 {
-        assert!(
-            !tree.contains(&format!("large{}.dat", i)),
-            "expected large{}.dat to be dropped",
-            i
-        );
-    }
-
-    // Verify correct count (should have small + medium files)
-    let files: Vec<&str> = tree.lines().collect();
-    let small_count: usize = files.iter().filter(|f| f.starts_with("small")).count();
-    let medium_count: usize = files
-        .iter()
-        .copied()
-        .filter(|f| f.starts_with("medium"))
-        .count();
-    let large_count: usize = files
-        .iter()
-        .copied()
-        .filter(|f| f.starts_with("large"))
-        .count();
-
-    assert_eq!(
-        small_count, 50,
-        "expected exactly 50 small files, got {}",
-        small_count
-    );
-    assert_eq!(
-        medium_count, 20,
-        "expected exactly 20 medium files, got {}",
-        medium_count
-    );
-    assert_eq!(
-        large_count, 0,
-        "expected no large files, got {}",
-        large_count
-    );
-}
-
-#[test]
-fn max_blob_size_fallback_behavior() {
-    // Create a fresh repository without the usual init_repo to avoid complexity
-    let repo = tempfile::TempDir::new().unwrap();
-    let repo_path = repo.path();
-
-    // Initialize git repo
-    let (c, _o, e) = run_git(&repo_path, &["init"]);
-    assert_eq!(c, 0, "git init failed: {}", e);
-    run_git(&repo_path, &["config", "user.name", "A U Thor"]).0;
-    run_git(
-        &repo_path,
-        &["config", "user.email", "a.u.thor@example.com"],
-    )
-    .0;
-
-    // Create just one simple test file
-    write_file(&repo_path, "test.txt", "hello"); // 5 bytes
-
-    run_git(&repo_path, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo_path, &["commit", "-q", "-m", "add test file"]).0,
-        0
-    );
-
-    // Verify file exists before filtering
-    let (_c0, tree0, _e0) = run_git(&repo_path, &["ls-tree", "-r", "--name-only", "HEAD"]);
-    assert!(
-        tree0.contains("test.txt"),
-        "test.txt should exist before filtering: {}",
-        tree0
-    );
-
-    // Test with large threshold - should definitely keep the file
-    let (_c, _o, _e) = run_tool(&repo_path, |o| {
-        // o.source = repo_path.to_path_buf();
-        // o.target = repo_path.to_path_buf();
-        o.max_blob_size = Some(1000); // 1KB threshold, 5 byte file should remain
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo_path,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    println!("After filtering: {}", tree);
-    assert!(
-        tree.contains("test.txt"),
-        "expected test.txt to remain (5 bytes < 1000)"
-    );
-}
-
-#[test]
-fn max_blob_size_no_git_objects() {
-    // Create a truly empty repository (no initial commit with files)
-    let repo = tempfile::TempDir::new().unwrap();
-    let repo_path = repo.path();
-
-    // Initialize git repo
-    let (c, _o, e) = run_git(repo_path, &["init"]);
-    assert_eq!(c, 0, "git init failed: {}", e);
-
-    // Configure git user
-    run_git(repo_path, &["config", "user.name", "test"]).0;
-    run_git(repo_path, &["config", "user.email", "test@example.com"]).0;
-
-    // Create empty commits (no blob objects)
-    run_git(
-        repo_path,
-        &["commit", "--allow-empty", "-q", "-m", "empty commit 1"],
-    )
-    .0;
-    run_git(
-        repo_path,
-        &["commit", "--allow-empty", "-q", "-m", "empty commit 2"],
-    )
-    .0;
-
-    // Should handle empty repositories gracefully
-    let (_c, _o, _e) = run_tool(repo_path, |o| {
-        // o.source = repo_path.to_path_buf();
-        // o.target = repo_path.to_path_buf();
-        o.max_blob_size = Some(1000);
-    });
-
-    // Should complete without errors even with no blobs
-    let (_c2, tree, _e2) = run_git(
-        repo_path,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        tree.is_empty(),
-        "expected no files in empty commit repository"
-    );
-}
-
-#[test]
-fn max_blob_size_corrupted_git_output() {
-    let repo = init_repo();
-
-    // Create a normal repository first
-    write_file(&repo, "test.txt", "test content");
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add test file"]).0,
-        0
-    );
-
-    // Test that the system can handle unexpected git output formats
-    // This simulates what happens when git returns unexpected output
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(5);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Should still work correctly despite any internal issues
-    assert!(
-        !tree.contains("test.txt"),
-        "expected test.txt to be dropped due to size"
-    );
-}
-
-#[test]
-fn max_blob_size_extreme_threshold_values() {
-    let repo = init_repo();
-
-    // Create test files of various sizes
-    write_file(&repo, "tiny.txt", "x"); // 1 byte
-    write_file(&repo, "small.txt", &"x".repeat(100)); // 100 bytes
-    write_file(&repo, "medium.txt", &"x".repeat(10000)); // 10KB
-    write_file(&repo, "large.txt", &"x".repeat(100000)); // 100KB
-
-    run_git(&repo, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo, &["commit", "-q", "-m", "add various sized files"]).0,
-        0
-    );
-
-    // Test with extremely small threshold (should drop almost everything)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1);
-    });
-    let (_c2, tree1, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        !tree1.contains("small.txt"),
-        "expected small.txt to be dropped with threshold 1"
-    );
-    assert!(
-        !tree1.contains("medium.txt"),
-        "expected medium.txt to be dropped with threshold 1"
-    );
-    assert!(
-        !tree1.contains("large.txt"),
-        "expected large.txt to be dropped with threshold 1"
-    );
-    // Only tiny.txt (1 byte) should remain
-    assert!(
-        tree1.contains("tiny.txt"),
-        "expected tiny.txt to remain with threshold 1"
-    );
-
-    // Create a fresh repository for the large threshold test
-    let repo2 = init_repo();
-    write_file(&repo2, "tiny.txt", "x"); // 1 byte
-    write_file(&repo2, "small.txt", &"x".repeat(100)); // 100 bytes
-    write_file(&repo2, "medium.txt", &"x".repeat(10000)); // 10KB
-    write_file(&repo2, "large.txt", &"x".repeat(100000)); // 100KB
-
-    run_git(&repo2, &["add", "."]).0;
-    assert_eq!(
-        run_git(&repo2, &["commit", "-q", "-m", "add various sized files"]).0,
-        0
-    );
-
-    // Test with extremely large threshold (should keep everything)
-    let (_c, _o, _e) = run_tool(&repo2, |o| {
-        o.max_blob_size = Some(1000000);
-    });
-    let (_c2, tree2, _e2) = run_git(
-        &repo2,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        tree2.contains("tiny.txt"),
-        "expected tiny.txt to remain with large threshold"
-    );
-    assert!(
-        tree2.contains("small.txt"),
-        "expected small.txt to remain with large threshold"
-    );
-    assert!(
-        tree2.contains("medium.txt"),
-        "expected medium.txt to remain with large threshold"
-    );
-    assert!(
-        tree2.contains("large.txt"),
-        "expected large.txt to remain with large threshold"
-    );
-}
-
-#[test]
-fn max_blob_size_normal_processing() {
-    let repo = init_repo();
-
-    // Create files that would trigger various fallback scenarios
-    std::fs::write(repo.join("normal.txt"), b"normal content").unwrap();
-    std::fs::write(repo.join("empty.txt"), b"").unwrap();
-    std::fs::write(repo.join("binary.dat"), vec![0u8; 1000]).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "add test files for fallback"]);
-
-    // Test with batch processing enabled (normal case)
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(500);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Should work normally with batch processing
-    assert!(
-        tree.contains("normal.txt"),
-        "expected normal.txt to remain (smaller than limit): {}",
-        tree
-    );
-    assert!(
-        tree.contains("empty.txt"),
-        "expected empty.txt to remain (zero bytes): {}",
-        tree
-    );
-    assert!(
-        !tree.contains("binary.dat"),
-        "expected binary.dat to be dropped (larger than limit): {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_zero_threshold() {
-    let repo = init_repo();
-
-    // Create files that would trigger various fallback scenarios
-    std::fs::write(repo.join("normal.txt"), b"normal content").unwrap();
-    std::fs::write(repo.join("empty.txt"), b"").unwrap();
-    std::fs::write(repo.join("binary.dat"), vec![0u8; 1000]).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "add test files for zero threshold test"],
-    );
-
-    // Test with zero threshold (edge case - should drop everything)
-    let (_c, tree, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(0);
-    });
-
-    assert!(
-        !tree.contains("normal.txt"),
-        "expected normal.txt to be dropped (zero threshold)"
-    );
-    assert!(
-        !tree.contains("empty.txt"),
-        "expected empty.txt to be dropped (zero threshold)"
-    );
-    assert!(
-        !tree.contains("binary.dat"),
-        "expected binary.dat to be dropped (zero threshold)"
-    );
-}
-
-#[test]
-fn max_blob_size_edge_case_handling() {
-    let repo = init_repo();
-
-    // Create a normal file first
-    std::fs::write(repo.join("normal.txt"), b"normal content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "add normal file"]);
-
-    // Test that normal processing works even with potential edge cases
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(100);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        tree.contains("normal.txt"),
-        "expected normal.txt to be processed correctly: {}",
-        tree
-    );
-}
-
-#[test]
-fn max_blob_size_memory_management() {
-    let repo = init_repo();
-
-    // Create many small files to test memory management
-    for i in 0..20 {
-        let content = format!("small file {} content", i);
-        std::fs::write(repo.join(format!("small_{}.txt", i)), content).unwrap();
-    }
-
-    // Create one large file
-    std::fs::write(repo.join("large.bin"), vec![0u8; 2000]).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "add files for memory efficiency test"],
-    );
-
-    // Test with small threshold - should handle memory efficiently
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(100);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // All small files should remain
-    for i in 0..20 {
-        assert!(
-            tree.contains(&format!("small_{}.txt", i)),
-            "expected small_{}.txt to remain (smaller than limit)",
-            i
-        );
-    }
-
-    // Large file should be dropped
-    assert!(
-        !tree.contains("large.bin"),
-        "expected large.bin to be dropped (larger than limit)"
-    );
-}
-
-#[test]
-fn max_blob_size_precise_threshold_handling() {
-    let repo = init_repo();
-
-    // Test exact threshold boundaries
-    std::fs::write(repo.join("exactly_100_bytes.txt"), b"a".repeat(100)).unwrap();
-    std::fs::write(repo.join("exactly_101_bytes.txt"), b"b".repeat(101)).unwrap();
-    std::fs::write(repo.join("just_under_100.txt"), b"c".repeat(99)).unwrap();
-    std::fs::write(repo.join("just_over_100.txt"), b"d".repeat(101)).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "add boundary test files"]);
-
-    // Test with 100 byte threshold
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(100);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Files <= 100 bytes should remain
-    assert!(
-        tree.contains("exactly_100_bytes.txt"),
-        "expected exactly_100_bytes.txt to remain (equal to limit)"
-    );
-    assert!(
-        tree.contains("just_under_100.txt"),
-        "expected just_under_100.txt to remain (under limit)"
-    );
-
-    // Files > 100 bytes should be dropped
-    assert!(
-        !tree.contains("exactly_101_bytes.txt"),
-        "expected exactly_101_bytes.txt to be dropped (over limit)"
-    );
-    assert!(
-        !tree.contains("just_over_100.txt"),
-        "expected just_over_100.txt to be dropped (over limit)"
-    );
-}
-
-// ===== PHASE 2.1: ERROR HANDLING TESTS =====
-
-#[test]
-fn error_handling_invalid_source_repository() {
-    let temp_dir = tempfile::TempDir::new().unwrap();
-    let invalid_repo = temp_dir.path().join("nonexistent");
-
-    // Test with non-existent source repository
-    let opts = fr::Options {
-        source: invalid_repo.clone(),
-        target: temp_dir.path().to_path_buf(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_err(),
-        "expected error for invalid source repository"
-    );
-
-    let error = result.err().unwrap();
-    let error_msg = format!("{:?}", error);
-    assert!(
-        error_msg.contains("not a git repo") || error_msg.contains("failed"),
-        "expected git repository error, got: {}",
-        error_msg
-    );
-}
-
-#[test]
-fn error_handling_invalid_target_repository() {
-    let repo = init_repo();
-    let temp_dir = tempfile::TempDir::new().unwrap();
-    let invalid_target = temp_dir
-        .path()
-        .join("nonexistent")
-        .join("nested")
-        .join("path");
-
-    // Test with invalid target path (parent doesn't exist)
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: invalid_target,
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_err(),
-        "expected error for invalid target repository"
-    );
-}
-
-#[test]
-fn error_handling_nonexistent_replace_message_file() {
-    let repo = init_repo();
-    let nonexistent_file = repo.join("nonexistent_replacements.txt");
-
-    // Test with non-existent replace-message file
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        replace_message_file: Some(nonexistent_file),
-        ..Default::default()
-    };
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_err(),
-        "expected error for non-existent replace-message file"
-    );
-
-    let error = result.err().unwrap();
-    let error_msg = format!("{:?}", error);
-    assert!(
-        error_msg.contains("replace-message") || error_msg.contains("failed to read"),
-        "expected replace-message file error, got: {}",
-        error_msg
-    );
-}
-
-#[test]
-fn error_handling_invalid_sha_format_in_strip_blobs() {
-    let repo = init_repo();
-    let invalid_sha_file = repo.join("invalid_shas.txt");
-
-    // Write invalid SHA formats (not 40 hex chars)
-    std::fs::write(&invalid_sha_file, "invalid123\nnotahash\nshort\n").unwrap();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "test commit"]);
-
-    // Test with invalid SHA formats
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        strip_blobs_with_ids: Some(invalid_sha_file),
-        ..Default::default()
-    };
-
-    // This should not crash but should handle invalid SHA gracefully
-    let _result = fr::run(&opts);
-    // The tool should either succeed (ignoring invalid SHAs) or fail gracefully
-    // We just want to ensure it doesn't panic
-}
-
-#[test]
-fn path_rename_with_identical_paths() {
-    let repo = init_repo();
-
-    // Test with invalid path rename format (missing colon)
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        path_renames: vec![(b"invalidformat".to_vec(), b"invalidformat".to_vec())], // Should be "old:new"
-        ..Default::default()
-    };
-
-    let _result = fr::run(&opts);
-    // This should handle the invalid format gracefully without panicking
-}
-
-#[test]
-fn error_handling_invalid_max_blob_size_values() {
-    let repo = init_repo();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "test commit"]);
-
-    // Test with various problematic max-blob-size values
-    let test_cases = vec![
-        Some(0),          // Zero should be handled gracefully
-        Some(1),          // Very small size
-        Some(usize::MAX), // Very large size
-    ];
-
-    for max_size in test_cases {
-        let opts = fr::Options {
-            source: repo.clone(),
-            target: repo.clone(),
-            refs: vec!["--all".to_string()],
-            max_blob_size: max_size,
-            ..Default::default()
-        };
-
-        // These should not panic, even with extreme values
-        let _result = fr::run(&opts);
-        // We don't care about success/failure, just that it doesn't panic
-    }
-}
-
-#[test]
-fn error_handling_malformed_utf8_content() {
-    let repo = init_repo();
-
-    // Create a file with malformed UTF-8 content
-    let malformed_utf8 = vec![
-        0x66, 0x69, 0x6c, 0x65, // "file" in valid UTF-8
-        0x80, 0x81, // Invalid UTF-8 sequence
-        0x2e, 0x74, 0x78, 0x74, // ".txt" in valid UTF-8
-    ];
-
-    std::fs::write(repo.join("test.bin"), malformed_utf8).unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "add malformed utf8"]);
-
-    // Test that malformed UTF-8 doesn't cause crashes
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    let _result = fr::run(&opts);
-    // Should handle malformed UTF-8 gracefully without panicking
-}
-
-#[test]
-fn error_handling_permission_denied_simulation() {
-    let repo = init_repo();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "test commit"]);
-
-    // Test with a read-only directory to simulate permission issues
-    let readonly_dir = repo.join("readonly");
-    std::fs::create_dir_all(&readonly_dir).unwrap();
-
-    // Make directory read-only (this might not work on all systems, but shouldn't crash)
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let mut perms = std::fs::metadata(&readonly_dir).unwrap().permissions();
-        perms.set_readonly(true);
-        std::fs::set_permissions(&readonly_dir, perms).unwrap();
-    }
-
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: readonly_dir.clone(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    let _result = fr::run(&opts);
-    // Should handle permission errors gracefully
-}
-
-#[test]
-fn error_handling_empty_replace_text_file() {
-    let repo = init_repo();
-    let empty_file = repo.join("empty_replacements.txt");
-
-    // Create an empty replacement file
-    std::fs::write(&empty_file, "").unwrap();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "test commit"]);
-
-    // Test with empty replace-text file
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        replace_text_file: Some(empty_file),
-        ..Default::default()
-    };
-
-    // Should handle empty file gracefully
-    let _result = fr::run(&opts);
-    // Empty file should not cause crashes
-}
-
-#[test]
-fn error_handling_corrupted_git_repository() {
-    let temp_dir = tempfile::TempDir::new().unwrap();
-    let repo_path = temp_dir.path();
-
-    // Initialize a normal git repository
-    run_git(&repo_path, &["init"]);
-
-    // Corrupt the git repository by removing essential files
-    let git_dir = repo_path.join(".git");
-    let objects_dir = git_dir.join("objects");
-    if objects_dir.exists() {
-        std::fs::remove_dir_all(&objects_dir).unwrap();
-    }
-
-    // Test with corrupted repository
-    let opts = fr::Options {
-        source: repo_path.to_path_buf(),
-        target: repo_path.to_path_buf(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    let result = fr::run(&opts);
-    assert!(result.is_err(), "expected error for corrupted repository");
-
-    let error = result.err().unwrap();
-    let error_msg = format!("{:?}", error);
-    assert!(
-        error_msg.contains("not a git repo") || error_msg.contains("failed"),
-        "expected repository corruption error, got: {}",
-        error_msg
-    );
-}
-
-#[test]
-fn error_handling_extremely_long_paths() {
-    let repo = init_repo();
-
-    // Create a file with an extremely long path name
-    let long_filename = "a".repeat(200); // 200 character filename
-    let long_path = repo.join(&long_filename);
-
-    std::fs::write(&long_path, "content").unwrap();
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "add long filename"]);
-
-    // Test with extremely long paths
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        paths: vec![long_filename.into_bytes()],
-        ..Default::default()
-    };
-
-    // Should handle long paths gracefully
-    let _result = fr::run(&opts);
-    // Long paths should not cause crashes
-}
-
-// === PHASE 2.2: MEMORY MANAGEMENT TESTS ===
-
-#[test]
-fn memory_management_large_blob_cache() {
-    let repo = init_repo();
-
-    // Create many blobs of varying sizes to test cache management
-    let mut blob_hashes = Vec::new();
-    for i in 0..1000 {
-        let content = if i % 10 == 0 {
-            // Some large blobs
-            "x".repeat(5000)
-        } else {
-            // Mostly small blobs
-            format!("content {}", i)
-        };
-
-        let path = format!("file_{}.txt", i);
-        std::fs::write(repo.join(&path), &content).unwrap();
-        run_git(&repo, &["add", &path]);
-
-        // Store blob hash for verification
-        let (_c, output, _e) = run_git(&repo, &["hash-object", "--stdin"]);
-        blob_hashes.push(output.trim().to_string());
-    }
-
-    run_git(&repo, &["commit", "-m", "add many files"]);
-
-    // Test with a size threshold that will filter many blobs
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Should filter large blobs while managing memory efficiently
-    let files: Vec<&str> = tree.split_whitespace().collect();
-    assert!(
-        files.len() < 1000,
-        "expected some files to be filtered by size"
-    );
-
-    // Verify no memory issues occurred (test would crash on memory errors)
-}
-
-#[test]
-fn memory_management_concurrent_operations() {
-    let repo = init_repo();
-
-    // Create multiple files to simulate concurrent access patterns
-    for i in 0..100 {
-        let content = match i % 4 {
-            0 => "x".repeat(2000),      // Large
-            1 => "x".repeat(500),       // Medium
-            2 => "x".repeat(100),       // Small
-            _ => "minimal".to_string(), // Tiny
-        };
-
-        let path = format!("concurrent_{}.txt", i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-    run_git(&repo, &["commit", "-m", "concurrent test"]);
-
-    // First, check the baseline (no filtering)
-    let (_c0, tree0, _e0) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files0: Vec<&str> = tree0.split_whitespace().collect();
-    let baseline_count = files0.len();
-    assert!(
-        baseline_count >= 100,
-        "baseline should have at least 100 files, got {}",
-        baseline_count
-    );
-
-    // Run multiple filter operations in sequence to test memory cleanup
-    for threshold in &[100, 500, 1000, 2000] {
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(*threshold);
-        });
-
-        // Each operation should clean up properly without memory leaks
-        let (_c2, tree, _e2) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-
-        // For memory management testing, we care more about the process completing successfully
-        // than exact file counts, since the main goal is to test that memory is cleaned up
-        // between operations and that the tool doesn't crash or leak memory
-        assert!(
-            files.len() <= baseline_count,
-            "file count should not exceed original count for threshold {}: {} vs {}",
-            threshold,
-            files.len(),
-            baseline_count
-        );
-
-        // Ensure filtering is actually working by verifying that files are being removed
-        // at lower thresholds (but allow for implementation-specific behavior)
-        if *threshold < 2000 {
-            assert!(
-                files.len() < baseline_count,
-                "filtering should remove some files at threshold {}: {} vs {}",
-                threshold,
-                files.len(),
-                baseline_count
-            );
-        }
-    }
-
-    // Final verification that all operations completed without memory issues
-}
-
-#[test]
-fn memory_management_path_filtering_memory() {
-    let repo = init_repo();
-
-    // Create many files in nested directories to test path filtering memory usage
-    for dir in 0..10 {
-        let dir_path = repo.join(format!("dir_{}", dir));
-        std::fs::create_dir_all(&dir_path).unwrap();
-
-        for file in 0..100 {
-            let content = format!("content for dir {} file {}", dir, file);
-            let file_path = dir_path.join(format!("file_{}.txt", file));
-            std::fs::write(file_path, content).unwrap();
-            run_git(&repo, &["add", &format!("dir_{}/file_{}.txt", dir, file)]);
-        }
-    }
-    run_git(&repo, &["commit", "-m", "nested directories"]);
-
-    // Test path filtering with many paths
-    let mut paths = Vec::new();
-    for dir in 0..5 {
-        paths.push(format!("dir_{}/", dir).into_bytes());
-    }
-
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        paths,
-        ..Default::default()
-    };
-
-    let _result = fr::run(&opts);
-
-    // Verify memory efficiency by checking operation completed
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    // Should have filtered files both by path and by size
-    assert!(
-        files.len() > 0 && files.len() < 1000,
-        "expected some files to remain after filtering"
-    );
-}
-
-#[test]
-fn memory_management_blob_size_precomputation_stress() {
-    let repo = init_repo();
-
-    // Create a repository with many commits and blobs to stress test precomputation
-    for commit in 0..20 {
-        for file in 0..50 {
-            let size = 100 + (commit * file * 10) % 4000; // Varying sizes
-            let content = "x".repeat(size);
-            let file_path = repo.join(format!("commit{}_file{}.txt", commit, file));
-            std::fs::write(file_path, content).unwrap();
-            run_git(
-                &repo,
-                &["add", &format!("commit{}_file{}.txt", commit, file)],
-            );
-        }
-        run_git(&repo, &["commit", "-m", &format!("commit {}", commit)]);
-    }
-
-    // Test precomputation with a size threshold that affects many blobs
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(1000);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Verify the operation completed without memory issues
-    let files: Vec<&str> = tree.split_whitespace().collect();
-    assert!(files.len() > 0, "expected some files to remain");
-
-    // The key test is that this doesn't crash due to memory pressure
-}
-
-#[test]
-fn memory_management_repeated_operations_same_repository() {
-    let repo = init_repo();
-
-    // Create test data
-    for i in 0..200 {
-        let size = 100 + (i * 15) % 3000;
-        let content = "x".repeat(size);
-        std::fs::write(repo.join(format!("file_{}.txt", i)), content).unwrap();
-        run_git(&repo, &["add", &format!("file_{}.txt", i)]);
-    }
-    run_git(&repo, &["commit", "-m", "initial commit"]);
-
-    // Run many consecutive operations to test for memory leaks
-    for iteration in 0..50 {
-        let threshold = 500 + (iteration * 50) % 2500;
-
-        let opts = fr::Options {
-            source: repo.clone(),
-            target: repo.clone(),
-            refs: vec!["--all".to_string()],
-            max_blob_size: Some(threshold),
-            ..Default::default()
-        };
-
-        let _result = fr::run(&opts);
-
-        // If there were memory leaks, this would eventually fail
-        if iteration % 10 == 0 {
-            let (_c, _tree, _e) = run_git(
-                &repo,
-                &[
-                    "-c",
-                    "core.quotepath=false",
-                    "ls-tree",
-                    "-r",
-                    "--name-only",
-                    "HEAD",
-                ],
-            );
-            // Test completed without memory issues
-        }
-    }
-
-    // Verify final state is consistent
-    let (_c, _tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    // All iterations completed successfully
-}
-
-#[test]
-fn memory_management_edge_case_empty_repositories() {
-    // Test with completely empty repositories
-    let temp_dir = tempfile::TempDir::new().unwrap();
-    let repo_path = temp_dir.path();
-
-    run_git(&repo_path, &["init"]);
-    // Create an empty commit
-    run_git(
-        &repo_path,
-        &["commit", "--allow-empty", "-m", "empty commit"],
-    );
-
-    let opts = fr::Options {
-        source: repo_path.to_path_buf(),
-        target: repo_path.to_path_buf(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(1000),
-        ..Default::default()
-    };
-
-    // Should handle empty repositories without memory issues
-    let _result = fr::run(&opts);
-
-    // Verify it completed without crashing
-    let (_c, tree, _e) = run_git(
-        &repo_path,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert_eq!(tree.trim(), "", "empty repository should have no files");
-}
-
-#[test]
-fn memory_management_extreme_path_depth() {
-    let repo = init_repo();
-
-    // Create deeply nested directory structure
-    let mut deep_path = repo.clone();
-    for level in 0..50 {
-        deep_path = deep_path.join(format!("level_{}", level));
-    }
-    std::fs::create_dir_all(deep_path.parent().unwrap()).unwrap();
-
-    // Create a file at maximum depth
-    let content = "content in deep path".repeat(10); // Make it sizable
-    std::fs::write(&deep_path, content).unwrap();
-
-    // Add the file (need to handle the long path)
-    let relative_path = deep_path.strip_prefix(&repo).unwrap();
-    let relative_str = relative_path.to_str().unwrap();
-    run_git(&repo, &["add", relative_str]);
-    run_git(&repo, &["commit", "-m", "deep path file"]);
-
-    // Test filtering with extreme path depth
-    let opts = fr::Options {
-        source: repo.clone(),
-        target: repo.clone(),
-        refs: vec!["--all".to_string()],
-        max_blob_size: Some(100),
-        ..Default::default()
-    };
-
-    // Should handle extreme path depth without memory issues
-    let _result = fr::run(&opts);
-
-    // Verify the file was processed
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    assert!(
-        !tree.trim().is_empty(),
-        "deep path file should be processed"
-    );
-}
-
-#[test]
-fn memory_management_unicode_path_heavy_load() {
-    let repo = init_repo();
-
-    // Create many files with Unicode characters to test memory usage with complex paths
-    let unicode_names = vec![
-        "café",
-        "naïve",
-        "résumé",
-        "séchage",
-        "noël",
-        "hiver",
-        "été",
-        "automne",
-        "北京",
-        "上海",
-        "广州",
-        "深圳",
-        "杭州",
-        "南京",
-        "成都",
-        "武汉",
-        "こんにちは",
-        "ありがとう",
-        "さようなら",
-        "おはよう",
-        "こんばんは",
-        "おやすみ",
-        "안녕하세요",
-        "감사합니다",
-        "안녕히 가세요",
-        "좋은 아침",
-        "좋은 저녁",
-    ];
-
-    for (i, base_name) in unicode_names.iter().cycle().take(200).enumerate() {
-        let path = format!("unicode_{}_{}.txt", i, base_name);
-        let content = format!("content for {} file {}", base_name, i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-    run_git(&repo, &["commit", "-m", "unicode heavy load"]);
-
-    // Test with size filtering on Unicode-heavy repository
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(100);
-    });
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-
-    // Should handle Unicode paths efficiently
-    let files: Vec<&str> = tree.split_whitespace().collect();
-    assert!(files.len() > 0, "unicode files should be processed");
-
-    // Verify Unicode characters are preserved correctly
-    let mut found_unicode_files = false;
-    for file in &files {
-        if file.contains("unicode_") {
-            found_unicode_files = true;
-            assert!(
-                file.contains("unicode_"),
-                "file should maintain unicode naming: {}",
-                file
-            );
-        }
-    }
-    assert!(
-        found_unicode_files,
-        "should find at least one unicode file, found: {:?}",
-        files
-    );
-}
-
-// ===== PHASE 2.3: CROSS-PLATFORM COMPATIBILITY TESTS =====
-
-#[test]
-fn cross_platform_windows_path_handling() {
-    let repo = init_repo();
-
-    // Test Windows-specific path patterns that might cause issues
-    // Note: Skip some reserved paths that Windows doesn't allow
-    let windows_paths = vec![
-        "file_with_backslash/path/test.txt",
-        "file/with/mixed/separators.txt",
-        "relative/path/file.txt",
-        "./hidden/file.txt",
-        "../parent/file.txt",
-    ];
-
-    for (i, path_str) in windows_paths.iter().enumerate() {
-        let content = format!("Windows path test file {} content", i);
-
-        // Create directory structure if needed
-        if let Some(parent) = std::path::Path::new(path_str).parent() {
-            if !parent.as_os_str().is_empty() {
-                if let Err(_e) = std::fs::create_dir_all(repo.join(parent)) {
-                    // Skip paths that can't be created on this platform
-                    continue;
-                }
-            }
-        }
-
-        if let Err(_e) = std::fs::write(repo.join(path_str), content) {
-            // Skip files that can't be created on this platform
-            continue;
-        }
-        run_git(&repo, &["add", path_str]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Windows path compatibility test"]);
-
-    // Test blob size filtering works correctly with Windows-style paths
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(50);
-    });
-
-    // Verify paths are handled correctly
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-    assert!(
-        files.len() > 0,
-        "Windows paths should be processed correctly"
-    );
-}
-
-#[test]
-fn cross_platform_case_sensitivity_handling() {
-    let repo = init_repo();
-
-    // Create files with different case variations
-    let test_files = vec![
-        "TestFile.txt",
-        "testfile.txt",
-        "TESTFILE.TXT",
-        "TestFile.TXT",
-        "subdir/File.txt",
-        "subdir/file.txt",
-    ];
-
-    for file_path in test_files {
-        if let Some(parent) = std::path::Path::new(file_path).parent() {
-            std::fs::create_dir_all(repo.join(parent)).unwrap();
-        }
-
-        std::fs::write(repo.join(file_path), format!("Content for {}", file_path)).unwrap();
-        run_git(&repo, &["add", file_path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Case sensitivity test"]);
-
-    // Test filtering handles case variations correctly
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(20);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    // Should preserve original case
-    assert!(
-        files.len() > 0,
-        "Case sensitivity should be handled correctly"
-    );
-}
-
-#[test]
-fn cross_platform_special_characters_in_paths() {
-    let repo = init_repo();
-
-    // Test various special characters that might behave differently across platforms
-    // Start with most universally supported characters
-    let special_paths = vec![
-        "file with spaces.txt",
-        "file-with-dashes.txt",
-        "file_with_underscores.txt",
-        "file.with.dots.txt",
-        "file123.txt", // simple alphanumeric as fallback
-    ];
-
-    let mut files_created = 0;
-    let mut successfully_created = Vec::new();
-
-    for (i, path_str) in special_paths.iter().enumerate() {
-        let content = format!("Special character test file {} content", i);
-
-        // Try to create the file, skip if it fails
-        match std::fs::write(repo.join(path_str), content) {
-            Ok(_) => {
-                let (code, _output, _error) = run_git(&repo, &["add", path_str]);
-                if code == 0 {
-                    files_created += 1;
-                    successfully_created.push(path_str.to_string());
-                }
-            }
-            Err(_e) => {
-                // Skip files that can't be created on this platform
-                continue;
-            }
-        }
-    }
-
-    // Only proceed with test if we created some files
-    if files_created > 0 {
-        run_git(&repo, &["commit", "-m", "Special characters in paths"]);
-
-        // Check what files are in the commit before running filter-repo
-        let (_c_pre, pre_tree, _e_pre) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let pre_files: Vec<&str> = pre_tree.split_whitespace().collect();
-
-        // Test blob size filtering with special characters
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(30);
-        });
-
-        let (_c2, tree, _e2) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-
-        assert!(
-            files.len() > 0,
-            "Special characters in paths should be handled correctly"
-        );
-
-        // Verify we can find our test files in the pre-filter state
-        let mut found_test_files = false;
-        for file in &pre_files {
-            for created_file in &successfully_created {
-                if file == created_file || created_file.contains(file) {
-                    found_test_files = true;
-                    break;
-                }
-            }
-        }
-
-        assert!(
-            found_test_files,
-            "should find at least one test file before filtering, created: {:?}, found: {:?}",
-            successfully_created, pre_files
-        );
-
-        // The main goal is that special characters don't cause crashes or panics
-        // Even if all files are filtered by size, the operation should complete successfully
-    } else {
-        // If no files could be created, create a simple test to verify the basic functionality
-        let simple_content = "Simple test content";
-        std::fs::write(repo.join("simple_file.txt"), simple_content).unwrap();
-        run_git(&repo, &["add", "simple_file.txt"]);
-        run_git(&repo, &["commit", "-m", "Simple fallback test"]);
-
-        // Test that blob size filtering works without crashing
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(5);
-        });
-
-        // If we get here without panicking, the test passes
-        println!("Special characters test fell back to simple file test");
-    }
-}
-
-#[test]
-fn cross_platform_unicode_normalization() {
-    let repo = init_repo();
-
-    // Test Unicode normalization forms (NFC, NFD, etc.)
-    let unicode_files = vec![
-        // Accented characters in different forms
-        "café.txt",                   // NFC form
-        "cafe\u{0301}.txt",           // NFD form (e + combining acute)
-        "naïve.txt",                  // NFC
-        "naive\u{0308}.txt",          // NFD
-        "résumé.txt",                 // NFC
-        "resume\u{0301}\u{0301}.txt", // NFD
-        // Other Unicode scripts
-        "русский.txt", // Cyrillic
-        "中文.txt",    // Chinese
-        "日本語.txt",  // Japanese
-        "العربية.txt", // Arabic
-        "한국어.txt",  // Korean
-        // Emoji and symbols
-        "🚀rocket.txt",
-        "⭐star.txt",
-        "♫music.txt",
-        // Zero-width characters
-        "zero\u{200B}width.txt", // Zero-width space
-        "zero\u{200C}join.txt",  // Zero-width non-joiner
-    ];
-
-    for (i, file_path) in unicode_files.iter().enumerate() {
-        let content = format!("Unicode normalization test file {} content", i);
-        std::fs::write(repo.join(file_path), content).unwrap();
-        run_git(&repo, &["add", file_path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Unicode normalization test"]);
-
-    // Test blob size filtering with Unicode normalization
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(50);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        files.len() > 0,
-        "Unicode files should be processed correctly"
-    );
-
-    // Verify Unicode characters are preserved
-    let mut found_unicode = false;
-    for file in files {
-        if !file.chars().all(|c| c.is_ascii()) {
-            found_unicode = true;
-            break;
-        }
-    }
-    assert!(
-        found_unicode,
-        "Should find Unicode characters in file names"
-    );
-}
-
-#[test]
-fn cross_platform_line_endings() {
-    let repo = init_repo();
-
-    // Create files with different line endings
-    let files_content = vec![
-        ("unix_line_endings.txt", "line1\nline2\nline3\n"),
-        ("windows_line_endings.txt", "line1\r\nline2\r\nline3\r\n"),
-        ("mixed_line_endings.txt", "line1\nline2\r\nline3\n"),
-        ("old_mac_line_endings.txt", "line1\rline2\rline3\r"),
-    ];
-
-    for (file_path, content) in files_content {
-        std::fs::write(repo.join(file_path), content).unwrap();
-        run_git(&repo, &["add", file_path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Line endings compatibility test"]);
-
-    // Test blob size filtering with different line endings
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(25);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        files.len() > 0,
-        "Files with different line endings should be processed"
-    );
-}
-
-#[test]
-fn cross_platform_file_permissions() {
-    let repo = init_repo();
-
-    // Create files with different permissions (where supported)
-    let test_files = vec!["normal_file.txt", "readonly_file.txt", "executable_file.sh"];
-
-    for file_path in test_files {
-        let content = format!("Test content for {}", file_path);
-        std::fs::write(repo.join(file_path), content).unwrap();
-
-        // Try to set different permissions (may not work on all platforms)
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let path = repo.join(file_path);
-            let mut perms = std::fs::metadata(&path).unwrap().permissions();
-
-            if file_path.contains("readonly") {
-                perms.set_readonly(true);
-            } else if file_path.contains("executable") {
-                perms.set_mode(0o755);
-            }
-
-            std::fs::set_permissions(&path, perms).unwrap();
-        }
-
-        run_git(&repo, &["add", file_path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "File permissions test"]);
-
-    // Test blob size filtering works regardless of file permissions
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(20);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        files.len() > 0,
-        "Files with different permissions should be processed"
-    );
-}
-
-#[test]
-fn cross_platform_long_file_names() {
-    let repo = init_repo();
-
-    // Test very long file names (approaching system limits)
-    let base_name = "a".repeat(200); // 200 character base name
-    let long_paths = vec![
-        format!("{}.txt", base_name),
-        format!("{}_{}.txt", base_name, "b".repeat(50)),
-        format!("long_directory_name_{}\\{}.txt", "x".repeat(100), base_name),
-    ];
-
-    for (i, path_str) in long_paths.iter().enumerate() {
-        let normalized_path = path_str.replace('\\', "/");
-        let content = format!("Long file name test {} content", i);
-
-        if let Some(parent) = std::path::Path::new(&normalized_path).parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(repo.join(parent)).unwrap();
-            }
-        }
-
-        std::fs::write(repo.join(&normalized_path), content).unwrap();
-        run_git(&repo, &["add", &normalized_path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Long file names test"]);
-
-    // Test blob size filtering with long file names
-    let (_c, _o, _e) = run_tool(&repo, |o| {
-        o.max_blob_size = Some(30);
-    });
-
-    let (_c2, tree, _e2) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        files.len() > 0,
-        "Long file names should be handled correctly"
-    );
-
-    // Verify file names are preserved
-    for file in files {
-        // README.md might be present from repo initialization, so check for our test files
-        if file.contains("long_file_name") {
-            assert!(
-                file.len() > 10,
-                "File names should maintain their length: {}",
-                file
-            );
-        }
-    }
-}
-
-// Phase 3.1: Performance Benchmark Tests
-// ===============================
-
-#[test]
-fn performance_large_repository_batch_optimization() {
-    let repo = init_repo();
-
-    // Create a large number of files to test batch optimization performance
-    let num_files = 1000;
-    let start_time = std::time::Instant::now();
-
-    for i in 0..num_files {
-        let content = match i % 5 {
-            0 => "large file content that exceeds typical blob size thresholds".repeat(100), // ~5KB
-            1 => "medium file content with moderate size".repeat(20), // ~800B
-            2 => "small file content".repeat(5),                      // ~100B
-            3 => "tiny content".to_string(),                          // ~12B
-            _ => "min".to_string(),                                   // ~3B
-        };
-
-        let path = format!("perf_test_file_{:04}.txt", i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-
-    let commit_time = start_time.elapsed();
-    run_git(
-        &repo,
-        &["commit", "-m", "Performance test: large repository"],
-    );
-    let setup_time = start_time.elapsed();
-
-    println!(
-        "Setup time: {:.2}s (commit: {:.2}s)",
-        setup_time.as_secs_f64(),
-        commit_time.as_secs_f64()
-    );
-
-    // Test baseline (no filtering)
-    let baseline_start = std::time::Instant::now();
-    let (_c0, tree0, _e0) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files0: Vec<&str> = tree0.split_whitespace().collect();
-    let baseline_count = files0.len();
-    let baseline_time = baseline_start.elapsed();
-
-    println!(
-        "Baseline: {} files in {:.2}ms",
-        baseline_count,
-        baseline_time.as_millis()
-    );
-    assert!(
-        baseline_count >= num_files,
-        "Should have at least {} files, got {}",
-        num_files,
-        baseline_count
-    );
-
-    // Test performance with different blob size thresholds
-    let thresholds = vec![10, 50, 100, 500, 1000, 5000];
-    let mut performance_metrics = Vec::new();
-
-    for threshold in thresholds {
-        let filter_start = std::time::Instant::now();
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(threshold);
-        });
-        let filter_time = filter_start.elapsed();
-
-        let verify_start = std::time::Instant::now();
-        let (_c2, tree, _e2) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-        let verify_time = verify_start.elapsed();
-
-        let filtered_count = files.len();
-        let filter_ratio = (baseline_count - filtered_count) as f64 / baseline_count as f64;
-
-        performance_metrics.push((threshold, filter_time, filtered_count, filter_ratio));
-
-        println!(
-            "Threshold {}: {} files ({:.1}% filtered) in {:.2}ms (verify: {:.2}ms)",
-            threshold,
-            filtered_count,
-            filter_ratio * 100.0,
-            filter_time.as_millis(),
-            verify_time.as_millis()
-        );
-
-        // Performance assertions - filtering should be reasonably fast even for large repositories
-        assert!(
-            filter_time.as_millis() < 5000,
-            "Filtering with threshold {} should complete within 5s, took {:.2}ms",
-            threshold,
-            filter_time.as_millis()
-        );
-
-        // Filtering should actually work (remove some files for reasonable thresholds)
-        if threshold < 5000 {
-            assert!(
-                filtered_count < baseline_count,
-                "Should filter some files for threshold {}: {} vs {}",
-                threshold,
-                filtered_count,
-                baseline_count
-            );
-        }
-    }
-
-    // Verify performance scaling - larger thresholds should not be significantly slower
-    if performance_metrics.len() >= 2 {
-        let fastest_time = performance_metrics
-            .iter()
-            .map(|&(_, t, _, _)| t)
-            .min()
-            .unwrap();
-        let slowest_time = performance_metrics
-            .iter()
-            .map(|&(_, t, _, _)| t)
-            .max()
-            .unwrap();
-        let time_ratio = slowest_time.as_millis() as f64 / fastest_time.as_millis() as f64;
-
-        println!("Performance ratio (slowest/fastest): {:.2}x", time_ratio);
-        assert!(
-            time_ratio < 10.0,
-            "Performance should not vary by more than 10x across thresholds, ratio: {:.2}x",
-            time_ratio
-        );
-    }
-
-    println!("Large repository performance test completed successfully");
-}
-
-#[test]
-fn performance_memory_usage_benchmark() {
-    let repo = init_repo();
-
-    // Create files that will test memory management with large blobs
-    let large_blob_count = 50;
-    let large_blob_size = 100_000; // 100KB each
-    let small_blob_count = 200;
-
-    // Create large blobs
-    for i in 0..large_blob_count {
-        let content = "x".repeat(large_blob_size);
-        let path = format!("large_blob_{:03}.dat", i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-
-    // Create small blobs
-    for i in 0..small_blob_count {
-        let content = format!("small blob {}", i);
-        let path = format!("small_blob_{:03}.txt", i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Memory benchmark test"]);
-
-    // Verify baseline
-    let (_c0, tree0, _e0) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files0: Vec<&str> = tree0.split_whitespace().collect();
-    let baseline_count = files0.len();
-    let expected_total = large_blob_count + small_blob_count;
-    assert!(
-        baseline_count >= expected_total,
-        "Should have at least {} files, got {}",
-        expected_total,
-        baseline_count
-    );
-
-    // Test memory usage with different filtering strategies
-    let test_cases = vec![
-        (Some(50_000), "filter large blobs"),
-        (Some(1000), "filter most blobs"),
-        (None, "no filtering"),
-    ];
-
-    for (max_size, description) in test_cases {
-        println!("Testing memory usage: {}", description);
-
-        // Create a fresh repository for each test case
-        let test_repo = init_repo();
-
-        // Create the same test files
-        for i in 0..large_blob_count {
-            let content = "x".repeat(large_blob_size);
-            let path = format!("large_blob_{:03}.dat", i);
-            std::fs::write(test_repo.join(&path), content).unwrap();
-            run_git(&test_repo, &["add", &path]);
-        }
-
-        for i in 0..small_blob_count {
-            let content = format!("small blob {}", i);
-            let path = format!("small_blob_{:03}.txt", i);
-            std::fs::write(test_repo.join(&path), content).unwrap();
-            run_git(&test_repo, &["add", &path]);
-        }
-
-        run_git(&test_repo, &["commit", "-m", "Memory benchmark test"]);
-
-        // Verify baseline for this test repository
-        let (_c0, tree0, _e0) = run_git(
-            &test_repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files0: Vec<&str> = tree0.split_whitespace().collect();
-        let test_baseline = files0.len();
-        let expected_total = large_blob_count + small_blob_count;
-        assert!(
-            test_baseline >= expected_total,
-            "Should have at least {} files, got {}",
-            expected_total,
-            test_baseline
-        );
-
-        // Measure memory before filtering
-        let start_time = std::time::Instant::now();
-        let (_c, _o, _e) = run_tool(&test_repo, |o| {
-            o.max_blob_size = max_size;
-        });
-        let elapsed = start_time.elapsed();
-
-        // Verify results
-        let (_c2, tree, _e2) = run_git(
-            &test_repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-        let filtered_count = files.len();
-
-        println!(
-            "  {} files remaining (from {}) in {:.2}ms",
-            filtered_count,
-            test_baseline,
-            elapsed.as_millis()
-        );
-
-        // Memory usage assertions - operations should complete in reasonable time
-        assert!(
-            elapsed.as_secs() < 30,
-            "Memory test should complete within 30s, took {:.2}s",
-            elapsed.as_secs()
-        );
-
-        // Verify filtering worked as expected
-        if let Some(size_limit) = max_size {
-            if size_limit == 50_000 {
-                // Should keep small blobs, filter large ones
-                assert!(
-                    filtered_count <= small_blob_count + 10,
-                    "Should mostly keep small blobs for 50KB limit: {} vs {}",
-                    filtered_count,
-                    small_blob_count
-                );
-            } else if size_limit == 1000 {
-                // Should filter large blobs (100KB) but keep small blobs (~15 bytes)
-                // We have 200 small blobs that should pass the 1KB filter
-                assert!(
-                    filtered_count >= small_blob_count - 10
-                        && filtered_count <= small_blob_count + 10,
-                    "Should keep small blobs for 1KB limit: {} vs expected ~{}",
-                    filtered_count,
-                    small_blob_count
-                );
-            }
-        } else {
-            // No filtering - should keep all files
-            assert!(
-                filtered_count == test_baseline,
-                "No filtering should keep all files: {} vs {}",
-                filtered_count,
-                test_baseline
-            );
-        }
-    }
-
-    println!("Memory usage benchmark completed successfully");
-}
-
-#[test]
-fn performance_cache_effectiveness() {
-    let repo = init_repo();
-
-    // Create a repository structure that benefits from caching
-    let num_commits = 20;
-    let files_per_commit = 10;
-
-    for commit_i in 0..num_commits {
-        for file_j in 0..files_per_commit {
-            let content = format!(
-                "Commit {} file {} content with varying size {}",
-                commit_i,
-                file_j,
-                "x".repeat((commit_i * 100 + file_j * 10) % 2000)
-            );
-            let path = format!("cache_test_commit_{:02}_file_{:02}.txt", commit_i, file_j);
-            std::fs::write(repo.join(&path), content).unwrap();
-            run_git(&repo, &["add", &path]);
-        }
-        run_git(
-            &repo,
-            &["commit", "-m", &format!("Cache test commit {}", commit_i)],
-        );
-    }
-
-    // Verify baseline
-    let (_c0, tree0, _e0) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files0: Vec<&str> = tree0.split_whitespace().collect();
-    let baseline_count = files0.len();
-    assert!(
-        baseline_count >= num_commits * files_per_commit,
-        "Should have at least {} files, got {}",
-        num_commits * files_per_commit,
-        baseline_count
-    );
-
-    // Test multiple filtering operations on the same repository
-    // This should benefit from any caching mechanisms
-    let num_iterations = 5;
-    let mut iteration_times = Vec::new();
-
-    for i in 0..num_iterations {
-        // Create a fresh repository state for each iteration
-        let fresh_repo = init_repo();
-
-        // Recreate the same repository structure
-        for commit_i in 0..num_commits {
-            for file_j in 0..files_per_commit {
-                let content = format!(
-                    "Commit {} file {} content with varying size {}",
-                    commit_i,
-                    file_j,
-                    "x".repeat((commit_i * 100 + file_j * 10) % 2000)
-                );
-                let path = format!("cache_test_commit_{:02}_file_{:02}.txt", commit_i, file_j);
-                std::fs::write(fresh_repo.join(&path), content).unwrap();
-                run_git(&fresh_repo, &["add", &path]);
-            }
-            run_git(
-                &fresh_repo,
-                &["commit", "-m", &format!("Cache test commit {}", commit_i)],
-            );
-        }
-
-        let start_time = std::time::Instant::now();
-        let (_c, _o, _e) = run_tool(&fresh_repo, |o| {
-            o.max_blob_size = Some(500);
-        });
-        let elapsed = start_time.elapsed();
-        iteration_times.push(elapsed);
-
-        let (_c2, tree, _e2) = run_git(
-            &fresh_repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-        let filtered_count = files.len();
-
-        println!(
-            "Iteration {}: {} files (from {}) in {:.2}ms",
-            i + 1,
-            filtered_count,
-            baseline_count,
-            elapsed.as_millis()
-        );
-    }
-
-    // Performance should not degrade significantly across iterations
-    // (in fact, it might improve due to caching)
-    if iteration_times.len() >= 3 {
-        let first_time = iteration_times[0];
-        let last_time = iteration_times[iteration_times.len() - 1];
-        let max_time = iteration_times.iter().max().unwrap();
-        let min_time = iteration_times.iter().min().unwrap();
-
-        println!(
-            "Cache performance - First: {:.2}ms, Last: {:.2}ms, Min: {:.2}ms, Max: {:.2}ms",
-            first_time.as_millis(),
-            last_time.as_millis(),
-            min_time.as_millis(),
-            max_time.as_millis()
-        );
-
-        // Performance should not degrade by more than 2x
-        let degradation_ratio = max_time.as_millis() as f64 / min_time.as_millis() as f64;
-        assert!(
-            degradation_ratio < 3.0,
-            "Performance should not degrade by more than 3x across iterations: {:.2}x",
-            degradation_ratio
-        );
-    }
-
-    println!("Cache effectiveness test completed successfully");
-}
-
-#[test]
-fn performance_scalability_with_blob_count() {
-    let repo = init_repo();
-
-    // Test scalability with increasing numbers of blobs
-    let blob_counts = vec![100, 500, 1000];
-    let mut scalability_metrics = Vec::new();
-
-    for &blob_count in &blob_counts {
-        println!("Testing scalability with {} blobs", blob_count);
-
-        // Create test files for this blob count
-        for i in 0..blob_count {
-            let content = format!(
-                "Scalability test blob {} with content size {}",
-                i,
-                "x".repeat((i % 1000) + 100)
-            );
-            let path = format!("scale_test_{:04}_count_{}.txt", i, blob_count);
-            std::fs::write(repo.join(&path), content).unwrap();
-            run_git(&repo, &["add", &path]);
-        }
-
-        run_git(
-            &repo,
-            &[
-                "commit",
-                "-m",
-                &format!("Scalability test with {} blobs", blob_count),
-            ],
-        );
-
-        // Test filtering performance
-        let start_time = std::time::Instant::now();
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(500);
-        });
-        let filter_time = start_time.elapsed();
-
-        let (_c2, tree, _e2) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-        let filtered_count = files.len();
-
-        scalability_metrics.push((blob_count, filter_time, filtered_count));
-
-        println!(
-            "  {} blobs: filtered to {} files in {:.2}ms",
-            blob_count,
-            filtered_count,
-            filter_time.as_millis()
-        );
-
-        // Performance should scale reasonably (not exponentially worse)
-        assert!(
-            filter_time.as_millis() < 10_000,
-            "Filtering should complete within 10s for {} blobs",
-            blob_count
-        );
-    }
-
-    // Analyze scalability - time should scale roughly linearly with blob count
-    if scalability_metrics.len() >= 2 {
-        println!("Scalability analysis:");
-        for (i, &(count, time, _filtered)) in scalability_metrics.iter().enumerate() {
-            if i > 0 {
-                let &(prev_count, prev_time, _) = &scalability_metrics[i - 1];
-                let count_ratio = count as f64 / prev_count as f64;
-                let time_ratio = time.as_millis() as f64 / prev_time.as_millis() as f64;
-                let efficiency = count_ratio / time_ratio;
-
-                println!(
-                    "  {} -> {} blobs: {:.1}x more blobs, {:.1}x more time, efficiency: {:.2}",
-                    prev_count, count, count_ratio, time_ratio, efficiency
-                );
-
-                // Efficiency should be reasonable (time shouldn't grow much faster than blob count)
-                assert!(
-                    efficiency > 0.3,
-                    "Efficiency should be reasonable: {:.2}",
-                    efficiency
-                );
-            }
-        }
-    }
-
-    println!("Scalability test completed successfully");
-}
-
-#[test]
-fn performance_batch_vs_individual_optimization() {
-    let repo = init_repo();
-
-    // Create a scenario that would benefit from batch processing
-    let blob_count = 500;
-    let mut blob_sizes = Vec::new();
-
-    for i in 0..blob_count {
-        let size = match i % 10 {
-            0 => 5000, // Large
-            1 => 2000, // Medium-large
-            2 => 1000, // Medium
-            3 => 500,  // Small-medium
-            4 => 100,  // Small
-            _ => 50,   // Very small
-        };
-        blob_sizes.push(size);
-
-        let content = "x".repeat(size);
-        let path = format!("batch_test_{:04}.txt", i);
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-
-    run_git(&repo, &["commit", "-m", "Batch optimization test"]);
-
-    // Verify baseline
-    let (_c0, tree0, _e0) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files0: Vec<&str> = tree0.split_whitespace().collect();
-    let baseline_count = files0.len();
-    assert!(
-        baseline_count >= blob_count,
-        "Should have at least {} files, got {}",
-        blob_count,
-        baseline_count
-    );
-
-    // Test multiple filtering thresholds to exercise batch processing
-    let thresholds = vec![10, 100, 1000, 5000];
-    let mut batch_times = Vec::new();
-
-    for &threshold in &thresholds {
-        let start_time = std::time::Instant::now();
-        let (_c, _o, _e) = run_tool(&repo, |o| {
-            o.max_blob_size = Some(threshold);
-        });
-        let elapsed = start_time.elapsed();
-        batch_times.push(elapsed);
-
-        let (_c2, tree, _e2) = run_git(
-            &repo,
-            &[
-                "-c",
-                "core.quotepath=false",
-                "ls-tree",
-                "-r",
-                "--name-only",
-                "HEAD",
-            ],
-        );
-        let files: Vec<&str> = tree.split_whitespace().collect();
-        let filtered_count = files.len();
-
-        println!(
-            "Threshold {}: {} files remaining in {:.2}ms",
-            threshold,
-            filtered_count,
-            elapsed.as_millis()
-        );
-
-        // Performance should be reasonable for batch processing
-        assert!(
-            elapsed.as_millis() < 5000,
-            "Batch processing should complete within 5s for threshold {}",
-            threshold
-        );
-    }
-
-    // Analyze performance characteristics
-    if batch_times.len() >= 2 {
-        let fastest = batch_times.iter().min().unwrap();
-        let slowest = batch_times.iter().max().unwrap();
-        let avg_time = batch_times.iter().sum::<std::time::Duration>() / batch_times.len() as u32;
-
-        println!(
-            "Batch performance - Fastest: {:.2}ms, Slowest: {:.2}ms, Average: {:.2}ms",
-            fastest.as_millis(),
-            slowest.as_millis(),
-            avg_time.as_millis()
-        );
-
-        // Performance should not vary wildly between thresholds
-        let variance_ratio = slowest.as_millis() as f64 / fastest.as_millis() as f64;
-        assert!(
-            variance_ratio < 5.0,
-            "Performance variance should be reasonable: {:.2}x",
-            variance_ratio
-        );
-    }
-
-    println!("Batch optimization test completed successfully");
-}
-
-// Phase 3.2: Complex multi-feature interaction tests
-#[test]
-fn multi_feature_blob_size_with_path_filtering() {
-    // Test combining --max-blob-size with --path filtering
-    let repo = init_repo();
-
-    // Create large files in different directories
-    std::fs::create_dir_all(repo.join("keep")).unwrap();
-    std::fs::create_dir_all(repo.join("filter")).unwrap();
-
-    std::fs::write(repo.join("keep/large_file.txt"), "x".repeat(2000)).unwrap();
-    std::fs::write(repo.join("filter/large_file.txt"), "x".repeat(2000)).unwrap();
-    std::fs::write(repo.join("keep/small_file.txt"), "small").unwrap();
-    std::fs::write(repo.join("filter/small_file.txt"), "small").unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(&repo, &["commit", "-m", "Add mixed size files"]);
-
-    // Run filter-repo with blob size limit AND path filtering
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.paths = vec![b"keep/".to_vec()];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(result.is_ok(), "Multi-feature filtering should succeed");
-
-    // Check that:
-    // 1. Only files in "keep/" directory remain
-    // 2. Large files in "keep/" are filtered out by size
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.starts_with("filter/")),
-        "Should filter out filter/ directory"
-    );
-    assert!(
-        files.contains(&"keep/small_file.txt"),
-        "Should keep small file in keep/ directory"
-    );
-    assert!(
-        !files.contains(&"keep/large_file.txt"),
-        "Should filter large file by size"
-    );
-    assert_eq!(
-        files.len(),
-        1,
-        "Should have exactly 1 file (small file in keep/)"
-    );
-}
-
-#[test]
-fn multi_feature_blob_size_with_path_rename() {
-    // Test combining --max-blob-size with --path-rename
-    let repo = init_repo();
-
-    // Create files with different sizes
-    std::fs::create_dir_all(repo.join("old_dir")).unwrap();
-    std::fs::create_dir_all(repo.join("other_dir")).unwrap();
-
-    std::fs::write(repo.join("old_dir/large_file.dat"), "x".repeat(1500)).unwrap();
-    std::fs::write(repo.join("old_dir/small_file.txt"), "small content").unwrap();
-    std::fs::write(repo.join("other_dir/medium_file.txt"), "x".repeat(800)).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add files for size and rename test"],
-    );
-
-    // Run filter-repo with blob size limit AND path rename
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.path_renames = vec![(b"old_dir/".to_vec(), b"new_dir/".to_vec())];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Multi-feature filtering with rename should succeed"
-    );
-
-    // Check that:
-    // 1. Files are renamed from old_dir/ to new_dir/
-    // 2. Large files are filtered out by size
-    // 3. Files in other directories are unaffected
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.starts_with("old_dir/")),
-        "Should rename old_dir/ to new_dir/"
-    );
-    assert!(
-        files.contains(&"new_dir/small_file.txt"),
-        "Should rename and keep small file"
-    );
-    assert!(
-        !files.contains(&"new_dir/large_file.dat"),
-        "Should filter large renamed file by size"
-    );
-    assert!(
-        files.contains(&"other_dir/medium_file.txt"),
-        "Should keep files in other directories"
-    );
-    assert!(
-        !files.iter().any(|f| f.contains("large_file.dat")),
-        "Large file should be filtered out regardless of rename"
-    );
-}
-
-#[test]
-fn multi_feature_blob_size_with_branch_rename() {
-    // Test combining --max-blob-size with --branch-rename
-    let repo = init_repo();
-
-    // Create files with different sizes
-    std::fs::write(repo.join("large_file.bin"), "x".repeat(2000)).unwrap();
-    std::fs::write(repo.join("small_file.txt"), "small").unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add files for size and branch rename test"],
-    );
-
-    // Create a branch
-    run_git(&repo, &["branch", "original-branch"]);
-
-    // Run filter-repo with blob size limit AND branch rename
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.branch_rename = Some((b"original-".to_vec(), b"renamed-".to_vec()));
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.refs = vec!["--all".to_string()];
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Multi-feature filtering with branch rename should succeed"
-    );
-
-    // Check that:
-    // 1. Branch is renamed
-    // 2. Large files are filtered out by size
-    let (_c, branches, _e) = run_git(&repo, &["branch", "-l"]);
-    assert!(
-        branches.contains("renamed-branch"),
-        "Branch should be renamed"
-    );
-    assert!(
-        !branches.contains("original-branch"),
-        "Original branch name should not exist"
-    );
-
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(files.contains(&"small_file.txt"), "Should keep small file");
-    assert!(
-        !files.contains(&"large_file.bin"),
-        "Should filter large file by size"
-    );
-}
-
-#[test]
-fn multi_feature_blob_size_with_tag_rename() {
-    // Test combining --max-blob-size with --tag-rename
-    let repo = init_repo();
-
-    // Create files with different sizes
-    std::fs::write(repo.join("large_file.dat"), "x".repeat(3000)).unwrap();
-    std::fs::write(repo.join("small_file.txt"), "small content").unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add files for size and tag rename test"],
-    );
-
-    // Create a tag
-    run_git(&repo, &["tag", "original-tag", "HEAD"]);
-
-    // Run filter-repo with blob size limit AND tag rename
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.tag_rename = Some((b"original-".to_vec(), b"renamed-".to_vec()));
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.refs = vec!["--all".to_string()];
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Multi-feature filtering with tag rename should succeed"
-    );
-
-    // Check that:
-    // 1. Tag is renamed
-    // 2. Large files are filtered out by size
-    let (_c, tags, _e) = run_git(&repo, &["tag", "-l"]);
-    assert!(tags.contains("renamed-tag"), "Tag should be renamed");
-    assert!(
-        !tags.contains("original-tag"),
-        "Original tag name should not exist"
-    );
-
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(files.contains(&"small_file.txt"), "Should keep small file");
-    assert!(
-        !files.contains(&"large_file.dat"),
-        "Should filter large file by size"
-    );
-}
-
-#[test]
-fn multi_feature_path_filtering_with_rename_and_size() {
-    // Test combining path filtering, path rename, and blob size limiting
-    let repo = init_repo();
-
-    // Create a complex directory structure with various file sizes
-    std::fs::create_dir_all(repo.join("src")).unwrap();
-    std::fs::create_dir_all(repo.join("docs")).unwrap();
-    std::fs::create_dir_all(repo.join("test")).unwrap();
-
-    std::fs::write(repo.join("src/large_module.rs"), "x".repeat(5000)).unwrap();
-    std::fs::write(repo.join("src/small_module.rs"), "small module").unwrap();
-    std::fs::write(repo.join("docs/large_doc.pdf"), "x".repeat(3000)).unwrap();
-    std::fs::write(repo.join("docs/small_doc.md"), "# Small doc").unwrap();
-    std::fs::write(repo.join("test/medium_test.rs"), "x".repeat(800)).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &[
-            "commit",
-            "-m",
-            "Add complex structure for multi-feature test",
-        ],
-    );
-
-    // Run filter-repo with multiple features:
-    // 1. Path filter to only include src/ and docs/
-    // 2. Path rename to change docs/ -> documentation/
-    // 3. Blob size limit to filter large files
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.paths = vec![b"src/".to_vec(), b"docs/".to_vec()];
-    opts.path_renames = vec![(b"docs/".to_vec(), b"documentation/".to_vec())];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Complex multi-feature filtering should succeed"
-    );
-
-    // Check that:
-    // 1. Only src/ and docs/ (renamed to documentation/) remain
-    // 2. docs/ is renamed to documentation/
-    // 3. Large files are filtered out by size
-    // 4. test/ directory is completely filtered out
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.starts_with("test/")),
-        "Should filter out test/ directory"
-    );
-    assert!(
-        files.contains(&"src/small_module.rs"),
-        "Should keep small src file"
-    );
-    assert!(
-        !files.contains(&"src/large_module.rs"),
-        "Should filter large src file by size"
-    );
-    assert!(
-        files.contains(&"documentation/small_doc.md"),
-        "Should rename docs/ and keep small file"
-    );
-    assert!(
-        !files.contains(&"documentation/large_doc.pdf"),
-        "Should filter large renamed file by size"
-    );
-    assert!(
-        !files.iter().any(|f| f.starts_with("docs/")),
-        "docs/ should be renamed to documentation/"
-    );
-
-    // Verify final file count is correct
-    assert_eq!(
-        files.len(),
-        2,
-        "Should have exactly 2 files: src/small_module.rs and documentation/small_doc.md"
-    );
-}
-
-#[test]
-fn multi_feature_invert_paths_with_size_filtering() {
-    // Test combining --invert-paths with --max-blob-size
-    let repo = init_repo();
-
-    // Create files with different sizes and locations
-    std::fs::create_dir_all(repo.join("exclude")).unwrap();
-    std::fs::create_dir_all(repo.join("include")).unwrap();
-
-    std::fs::write(repo.join("exclude/large_file.bin"), "x".repeat(2000)).unwrap();
-    std::fs::write(repo.join("exclude/small_file.txt"), "small").unwrap();
-    std::fs::write(repo.join("include/large_file.dat"), "x".repeat(1500)).unwrap();
-    std::fs::write(repo.join("include/small_file.txt"), "small content").unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add files for invert paths and size test"],
-    );
-
-    // Run filter-repo with:
-    // 1. Path filter to exclude/ directory
-    // 2. Invert paths to keep everything EXCEPT exclude/
-    // 3. Blob size limit to filter large files
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.paths = vec![b"exclude/".to_vec()];
-    opts.invert_paths = true;
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Invert paths with size filtering should succeed"
-    );
-
-    // Check that:
-    // 1. exclude/ directory is completely filtered out
-    // 2. Large files in include/ are filtered out by size
-    // 3. Small files in include/ are kept
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.starts_with("exclude/")),
-        "Should completely exclude exclude/ directory"
-    );
-    assert!(
-        files.contains(&"include/small_file.txt"),
-        "Should keep small file in include/"
-    );
-    assert!(
-        !files.contains(&"include/large_file.dat"),
-        "Should filter large file by size"
-    );
-    assert_eq!(
-        files.len(),
-        2,
-        "Should have exactly 2 files: README.md and include/small_file.txt"
-    );
-}
-
-#[test]
-fn multi_feature_complex_rename_chain() {
-    // Test complex rename operations with size filtering
-    let repo = init_repo();
-
-    // Create files in nested directories with various sizes
-    std::fs::create_dir_all(repo.join("old/v1")).unwrap();
-    std::fs::create_dir_all(repo.join("old/v2")).unwrap();
-
-    std::fs::write(repo.join("old/v1/large_file.cpp"), "x".repeat(4000)).unwrap();
-    std::fs::write(repo.join("old/v1/small_file.cpp"), "small").unwrap();
-    std::fs::write(repo.join("old/v2/medium_file.py"), "x".repeat(1200)).unwrap();
-    std::fs::write(repo.join("old/v2/tiny_file.py"), "tiny").unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &[
-            "commit",
-            "-m",
-            "Add nested structure for complex rename test",
-        ],
-    );
-
-    // Run filter-repo with multiple renames and size filtering:
-    // 1. old/ -> new/
-    // 2. v1/ -> version1/ (within new/)
-    // 3. v2/ -> version2/ (within new/)
-    // 4. Size limit to filter large files
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.path_renames = vec![
-        (b"old/".to_vec(), b"new/".to_vec()),
-        (b"new/v1/".to_vec(), b"new/version1/".to_vec()),
-        (b"new/v2/".to_vec(), b"new/version2/".to_vec()),
-    ];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Complex rename chain with size filtering should succeed"
-    );
-
-    // Check that:
-    // 1. All renames are applied in sequence
-    // 2. Large files are filtered out by size
-    // 3. Directory structure is preserved correctly
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.starts_with("old/")),
-        "old/ should be renamed to new/"
-    );
-    assert!(
-        files.contains(&"new/version1/small_file.cpp"),
-        "Should have renamed and kept small file"
-    );
-    assert!(
-        !files.contains(&"new/version1/large_file.cpp"),
-        "Should filter large file by size"
-    );
-    assert!(
-        files.contains(&"new/version2/tiny_file.py"),
-        "Should have renamed and kept tiny file"
-    );
-    assert!(
-        !files.contains(&"new/version2/medium_file.py"),
-        "Should filter medium file by size"
-    );
-
-    // Verify final structure (README.md is also present)
-    assert_eq!(
-        files.len(),
-        3,
-        "Should have exactly 3 files after renames and size filtering (including README.md)"
-    );
-}
-
-#[test]
-fn multi_feature_size_filter_with_special_paths() {
-    // Test size filtering with special path characters and Unicode
-    let repo = init_repo();
-
-    // Create files with special characters and various sizes
-    std::fs::create_dir_all(repo.join("path with spaces")).unwrap();
-    std::fs::create_dir_all(repo.join("unicode")).unwrap();
-    std::fs::create_dir_all(repo.join("regular")).unwrap();
-
-    std::fs::write(
-        repo.join("path with spaces/large file.txt"),
-        "x".repeat(2500),
-    )
-    .unwrap();
-    std::fs::write(repo.join("path with spaces/small file.txt"), "small").unwrap();
-    std::fs::write(repo.join("unicode/😀_large.dat"), "x".repeat(1800)).unwrap();
-    std::fs::write(repo.join("unicode/😀_small.dat"), "unicode small").unwrap();
-    std::fs::write(repo.join("regular/medium.txt"), "x".repeat(900)).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add special paths with size variation"],
-    );
-
-    // Run filter-repo with:
-    // 1. Size limit to filter large files
-    // 2. Path filter to include unicode/ and regular/ directories
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.paths = vec![b"unicode/".to_vec(), b"regular/".to_vec()];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Size filtering with special paths should succeed"
-    );
-
-    // Check that:
-    // 1. "path with spaces/" directory is filtered out
-    // 2. Large files in unicode/ and regular/ are filtered by size
-    // 3. Small files are kept
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        !files.iter().any(|f| f.contains("path with spaces")),
-        "Should filter out 'path with spaces' directory"
-    );
-
-    // Check for remaining files
-    let has_unicode_small = files.iter().any(|f| f.contains("😀_small.dat"));
-    let has_regular_medium = files.iter().any(|f| f.contains("regular/medium.txt"));
-
-    assert!(has_unicode_small, "Should keep unicode small file");
-    assert!(
-        has_regular_medium,
-        "Should keep regular medium file (under 1KB)"
-    );
-
-    // Should not have large files
-    assert!(
-        !files.iter().any(|f| f.contains("😀_large.dat")),
-        "Should filter unicode large file"
-    );
-
-    assert!(files.len() <= 2, "Should have at most 2 files");
-}
-
-#[test]
-fn multi_feature_empty_filtering_results() {
-    // Test edge case where multiple filters result in no files remaining
-    let repo = init_repo();
-
-    // Create only large files in directories that will be filtered
-    std::fs::create_dir_all(repo.join("will_be_filtered")).unwrap();
-    std::fs::create_dir_all(repo.join("also_filtered")).unwrap();
-
-    std::fs::write(
-        repo.join("will_be_filtered/large_file.bin"),
-        "x".repeat(3000),
-    )
-    .unwrap();
-    std::fs::write(repo.join("also_filtered/huge_file.dat"), "x".repeat(5000)).unwrap();
-
-    run_git(&repo, &["add", "."]);
-    run_git(
-        &repo,
-        &["commit", "-m", "Add files that will all be filtered"],
-    );
-
-    // Run filter-repo with:
-    // 1. Path filter to only include will_be_filtered/
-    // 2. Size limit that filters all files in that directory
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(1000);
-    opts.paths = vec![b"will_be_filtered/".to_vec()];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(
-        result.is_ok(),
-        "Filtering that results in no files should still succeed"
-    );
-
-    // Check that no files remain
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    assert!(
-        files.is_empty(),
-        "Should have no files remaining after filtering"
-    );
-
-    // But commit should still exist
-    let (_c, log, _e) = run_git(&repo, &["log", "--oneline"]);
-    assert!(
-        log.contains("Add files that will all be filtered"),
-        "Commit should exist but be empty"
-    );
-}
-
-#[test]
-fn multi_feature_performance_with_multiple_filters() {
-    // Test that performance is maintained when using multiple filters
-    let repo = init_repo();
-
-    // Create many files with various sizes and paths
-    for i in 0..100 {
-        let size = if i % 10 == 0 { 2000 } else { 100 }; // 10% large files
-        let content = "x".repeat(size);
-        let path = if i % 2 == 0 {
-            format!("keep/file_{}.txt", i)
-        } else {
-            format!("discard/file_{}.txt", i)
-        };
-
-        if let Some(parent) = std::path::Path::new(&path).parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(repo.join(parent)).unwrap();
-            }
-        }
-
-        std::fs::write(repo.join(&path), content).unwrap();
-        run_git(&repo, &["add", &path]);
-    }
-
-    run_git(
-        &repo,
-        &["commit", "-m", "Add many files for performance test"],
-    );
-
-    // Time the filtering operation
-    let start = std::time::Instant::now();
-
-    // Run filter-repo with multiple filters
-    let mut opts = fr::Options::default();
-    opts.max_blob_size = Some(500);
-    opts.paths = vec![b"keep/".to_vec()];
-    opts.path_renames = vec![(b"keep/".to_vec(), b"preserved/".to_vec())];
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    let duration = start.elapsed();
-
-    assert!(
-        result.is_ok(),
-        "Multi-filter performance test should succeed"
-    );
-
-    // Should complete in reasonable time (adjust threshold as needed)
-    assert!(
-        duration.as_secs() < 10,
-        "Multi-filter operation should complete quickly, took: {:?}",
-        duration
-    );
-
-    // Verify results are correct
-    let (_c, tree, _e) = run_git(
-        &repo,
-        &[
-            "-c",
-            "core.quotepath=false",
-            "ls-tree",
-            "-r",
-            "--name-only",
-            "HEAD",
-        ],
-    );
-    let files: Vec<&str> = tree.split_whitespace().collect();
-
-    // All files should be in preserved/ directory and small
-    for file in &files {
-        assert!(
-            file.starts_with("preserved/"),
-            "All files should be renamed to preserved/"
-        );
-        assert!(
-            !file.contains("discard/"),
-            "No files should remain from discard/"
-        );
-    }
-
-    // Should have filtered out large files and discard/ directory
-    assert!(
-        files.len() < 50,
-        "Should have significantly fewer files after filtering"
-    );
-}
-
-// Phase 4.1: Unit tests for remaining core modules
-#[test]
-fn unit_test_commit_message_processing() {
-    // Test that commit messages are processed correctly
-    let repo = init_repo();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "test.txt"]);
-    run_git(&repo, &["commit", "-m", "Original commit message"]);
-
-    // Test commit message replacement
-    let message_file = repo.join("message_replacements.txt");
-    std::fs::write(&message_file, "Original==>Replacement").unwrap();
-
-    let mut opts = fr::Options::default();
-    opts.replace_message_file = Some(message_file);
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-
-    let result = fr::run(&opts);
-    assert!(result.is_ok(), "Commit message replacement should succeed");
-
-    // Check that message was replaced
-    let (_c, log, _e) = run_git(&repo, &["log", "--oneline", "-1"]);
-    assert!(
-        log.contains("Replacement"),
-        "Commit message should be replaced"
-    );
-    assert!(
-        !log.contains("Original"),
-        "Original message should be replaced"
-    );
-}
-
-#[test]
-fn unit_test_tag_processing() {
-    // Test that tags are processed correctly
-    let repo = init_repo();
-
-    std::fs::write(repo.join("test.txt"), "test content").unwrap();
-    run_git(&repo, &["add", "test.txt"]);
-    run_git(&repo, &["commit", "-m", "Test commit for tags"]);
-
-    // Create lightweight and annotated tags
-    run_git(&repo, &["tag", "lightweight-tag"]);
-    run_git(
-        &repo,
-        &["tag", "-a", "annotated-tag", "-m", "Annotated tag message"],
-    );
-
-    let mut opts = fr::Options::default();
-    opts.tag_rename = Some((b"lightweight-".to_vec(), b"renamed-lightweight-".to_vec()));
-    opts.source = repo.clone();
-    opts.target = repo.clone();
-    opts.refs = vec!["--all".to_string()];
-
-    let result = fr::run(&opts);
-    assert!(result.is_ok(), "Tag processing should succeed");
-
-    // Check that lightweight tag was renamed
-    let (_c, tags, _e) = run_git(&repo, &["tag", "-l"]);
-    println!("Available tags: {:?}", tags);
-    let tags_list: Vec<&str> = tags.split('\n').collect();
-    assert!(
-        tags_list.contains(&"renamed-lightweight-tag"),
-        "Lightweight tag should be renamed"
-    );
-    assert!(
-        !tags_list.contains(&"lightweight-tag"),
-        "Original lightweight tag should not exist"
-    );
-    // Annotated tag should remain unchanged
-    assert!(
-        tags_list.contains(&"annotated-tag"),
-        "Annotated tag should remain"
-    );
-}
-
-#[test]
-fn unit_test_path_utilities() {
-    // Test path utility functions
-    use filter_repo_rs::pathutil;
-
-    // Test C-style dequoting on a string without outer quotes
-    let unquoted = b"test\npath\tab";
-    let dequoted = pathutil::dequote_c_style_bytes(unquoted);
-    assert_eq!(dequoted, b"test\npath\tab");
-
-    // Test unquoted input
-    let unquoted = b"regular_path";
-    let result = pathutil::dequote_c_style_bytes(unquoted);
-    assert_eq!(result, unquoted);
-
-    // Test empty input
-    let empty = b"";
-    let result = pathutil::dequote_c_style_bytes(empty);
-    assert_eq!(result, empty);
-}
-
-#[test]
-fn unit_test_git_utilities() {
-    // Test Git utility functions
-    let repo = init_repo();
-
-    // Test git show-ref functionality
-    std::fs::write(repo.join("test.txt"), "test").unwrap();
-    run_git(&repo, &["add", "test.txt"]);
-    run_git(&repo, &["commit", "-m", "Test commit"]);
-
-    let (_c, head_ref_out, _e) = run_git(&repo, &["symbolic-ref", "HEAD"]);
-    let head_ref = head_ref_out.trim();
-    let (_c, output, _e) = run_git(&repo, &["show-ref", head_ref]);
-    println!("show-ref output: {:?}", output);
-    assert!(!output.is_empty(), "show-ref should return HEAD reference");
-}
